refactor(fgo): split the 5.8k-line fgo.cpp into per-concern TUs
- fgo_internal.hpp: former anonymous-namespace helpers (signal
  eligibility, tracking-code tables, carrier prep, base-epoch matching)
  as inline functions in libgnss::fgo_internal, plus the GTSAM backend
  forward declaration
- fgo_problems.cpp: buildPseudorangeProblem / buildDoubleDifferenceProblem
- fgo_shadows.cpp: geometry-free slip, temporal-carrier-shadow,
  predicted-DDPR quality/bias diagnostic analyzers
- fgo.cpp keeps the optimize overloads and optimizeProblem (the native
  Eigen solver remains one function; splitting it is a future refactor)

No changes inside function bodies.
