+++ CMakeLists.txt
@@ -170,6 +170,8 @@ set(GNSS_RTK_RUNTIME_SOURCES
 set(GNSS_SOLVER_SOURCES
     src/algorithms/fgo_ddpr_gnc.cpp
     src/algorithms/fgo.cpp
+    src/algorithms/fgo_problems.cpp
+    src/algorithms/fgo_shadows.cpp
     src/algorithms/ppp_env_overrides.cpp
 )
