+++ src/algorithms/fgo.cpp
@@ -25,2614 +25,12 @@
 #include <string>
 #include <tuple>
 
+
+#include "fgo_internal.hpp"
+
 namespace libgnss {
 
-#ifdef GNSSPP_HAS_GTSAM
-// Implemented in fgo_gtsam_backend.cpp. Kept as a free function (rather than
-// another FGOProcessor member) so all GTSAM includes/types stay confined to
-// that translation unit; fgo.cpp itself never includes any GTSAM header.
-FGOProcessor::FGOResult optimizeProblemWithGtsam(
-    const FGOProcessor::FGOProblem& problem,
-    const FGOProcessor::FGOConfig& config,
-    FGOProcessor::FGOResult result);
-#endif
-
-namespace {
-
-bool isPrimaryFgoSignal(SignalType signal, bool use_multi_constellation) {
-    if (!use_multi_constellation) {
-        return signal == SignalType::GPS_L1CA;
-    }
-
-    switch (signal) {
-        case SignalType::GPS_L1CA:
-        case SignalType::GLO_L1CA:
-        case SignalType::GAL_E1:
-        case SignalType::BDS_B1I:
-        case SignalType::BDS_B1C:
-        case SignalType::QZS_L1CA:
-            return true;
-        default:
-            return false;
-    }
-}
-
-// Multi-frequency DD gate: primary-band signals are always eligible (same as
-// isPrimaryFgoSignal above). When use_multi_frequency_double_difference is
-// also enabled (and multi-constellation is on), secondary-band signals
-// (GPS L2C/L5, Galileo E5A/E5B/E6, BeiDou B2I/B3I/B2A, QZSS L2C/L5, ...) are
-// eligible too, keyed by the satellite's system via signal_policy. This is
-// the single choke point that previously restricted buildPseudorangeProblem /
-// buildDoubleDifferenceProblem to L1/E1/B1I-only DD factors.
-bool isEligibleFgoSignal(const SatelliteId& satellite,
-                         SignalType signal,
-                         bool use_multi_constellation,
-                         bool use_multi_frequency_double_difference) {
-    if (isPrimaryFgoSignal(signal, use_multi_constellation)) {
-        return true;
-    }
-    if (!use_multi_constellation || !use_multi_frequency_double_difference) {
-        return false;
-    }
-    return signal_policy::isSecondarySignal(satellite.system, signal);
-}
-
-// --- MF hygiene: per-receiver secondary-band observation-code alignment ---
-//
-// Port of the cssrlib/inuex35 (RTKLIB-style) signal-selection policy
-// (tightly-coupled-gnss-imu-fgo utils/sig_autodetect.py): each RECEIVER uses
-// exactly ONE RINEX observation code per (system, band) for the whole file,
-// so every satellite of a system contributes the same code per band and the
-// between-satellite (single) difference cancels the receiver-side inter-code
-// bias exactly. Without this, the RINEX reader's per-satellite first-nonzero
-// column selection MIXES codes across satellites within one band (observed on
-// tokyo1: rover GPS L2 = C2W for 2527 sat-epochs but C2L for the satellites/
-// epochs where C2W is empty; base BeiDou B2I = C7D vs rover C7I), leaving
-// uncancelled inter-code/DCB and carrier phase-alignment biases in the DD --
-// the measured multi-frequency FLOAT blowup. Satellites whose selected column
-// is a different code simply drop that band (like-vs-like or nothing),
-// exactly like cssrlib decoding only the chosen column.
-//
-// The preferred code per (system, SignalType) is chosen per receiver from the
-// codes actually observed in its data, ordered by an RTKLIB-style tracking-
-// code priority string (rtkcmn.c codepris), ties broken by observation count.
-// Applied to SECONDARY bands only: the primary band (L1/E1/B1I) single-
-// frequency path is measured-good and stays bit-identical.
-
-// RTKLIB-style per-(system, band) tracking-code priority (earlier = higher).
-const char* trackingCodePriorityString(GNSSSystem system, int band) {
-    switch (system) {
-        case GNSSSystem::GPS:
-            switch (band) {
-                case 1: return "CPYWMNSL";
-                case 2: return "PYWCMNDLSX";
-                case 5: return "IQX";
-                default: return "";
-            }
-        case GNSSSystem::GLONASS:
-            switch (band) {
-                case 1: return "PC";
-                case 2: return "PC";
-                case 3: return "IQX";
-                default: return "";
-            }
-        case GNSSSystem::Galileo:
-            switch (band) {
-                case 1: return "CABXZ";
-                case 5: return "IQX";
-                case 6: return "ABCXZ";
-                case 7: return "IQX";
-                case 8: return "IQX";
-                default: return "";
-            }
-        case GNSSSystem::BeiDou:
-            switch (band) {
-                case 1: return "DPX";
-                case 2: return "IQX";
-                case 5: return "DPX";
-                case 6: return "IQX";
-                case 7: return "IQXDZ";
-                case 8: return "DPX";
-                default: return "";
-            }
-        case GNSSSystem::QZSS:
-            switch (band) {
-                case 1: return "CLSXZ";
-                case 2: return "LSX";
-                case 5: return "IQXDPZ";
-                case 6: return "SXZ";
-                default: return "";
-            }
-        case GNSSSystem::NavIC:
-            switch (band) {
-                case 5: return "ABCX";
-                default: return "";
-            }
-        default:
-            return "";
-    }
-}
-
-int trackingCodePriority(GNSSSystem system, const std::string& obs_type) {
-    if (obs_type.size() < 3) {
-        return 100;
-    }
-    const int band = signal_policy::rinexBand(obs_type);
-    const char code = obs_type[2];
-    const char* priorities = trackingCodePriorityString(system, band);
-    for (int i = 0; priorities[i] != '\0'; ++i) {
-        if (priorities[i] == code) {
-            return i;
-        }
-    }
-    return 100;
-}
-
-// One preferred tracking-code char per (system, secondary SignalType),
-// derived from the receiver's own observations (single pre-scan pass).
-using SecondaryCodeTable = std::map<std::pair<GNSSSystem, SignalType>, char>;
-
-SecondaryCodeTable buildSecondaryCodeTable(
-    const std::vector<ObservationData>& epochs) {
-    struct CodeStats {
-        int priority = 100;
-        std::size_t count = 0;
-    };
-    std::map<std::pair<GNSSSystem, SignalType>, std::map<char, CodeStats>> stats;
-    for (const auto& epoch : epochs) {
-        for (const auto& observation : epoch.observations) {
-            if (!observation.valid || !observation.has_pseudorange) {
-                continue;
-            }
-            const GNSSSystem system = observation.satellite.system;
-            if (!signal_policy::isSecondarySignal(system, observation.signal)) {
-                continue;
-            }
-            const std::string& obs_type = observation.exactBiasObservationType();
-            if (obs_type.size() < 3) {
-                continue;
-            }
-            auto& entry = stats[{system, observation.signal}][obs_type[2]];
-            entry.priority = trackingCodePriority(system, obs_type);
-            ++entry.count;
-        }
-    }
-
-    SecondaryCodeTable table;
-    for (const auto& [key, codes] : stats) {
-        char best_code = '\0';
-        int best_priority = 101;
-        std::size_t best_count = 0;
-        for (const auto& [code, s] : codes) {
-            if (s.priority < best_priority ||
-                (s.priority == best_priority && s.count > best_count)) {
-                best_code = code;
-                best_priority = s.priority;
-                best_count = s.count;
-            }
-        }
-        if (best_code != '\0') {
-            table[key] = best_code;
-        }
-    }
-    return table;
-}
-
-// True when the observation either is primary-band (never filtered here) or
-// carries the receiver's preferred tracking code for its (system, signal).
-bool passesSecondaryCodeAlignment(const Observation& observation,
-                                  const SecondaryCodeTable* table) {
-    if (table == nullptr) {
-        return true;
-    }
-    const GNSSSystem system = observation.satellite.system;
-    if (!signal_policy::isSecondarySignal(system, observation.signal)) {
-        return true;
-    }
-    const auto it = table->find({system, observation.signal});
-    if (it == table->end()) {
-        return true;
-    }
-    const std::string& obs_type = observation.exactBiasObservationType();
-    return obs_type.size() >= 3 && obs_type[2] == it->second;
-}
-
-GNSSSystem clockBiasGroup(GNSSSystem system) {
-    switch (system) {
-        case GNSSSystem::GPS:
-        case GNSSSystem::QZSS:
-            return GNSSSystem::GPS;
-        case GNSSSystem::Galileo:
-        case GNSSSystem::BeiDou:
-        case GNSSSystem::GLONASS:
-        case GNSSSystem::NavIC:
-            return system;
-        default:
-            return GNSSSystem::UNKNOWN;
-    }
-}
-
-bool usesSeparateClockBias(GNSSSystem group) {
-    return group != GNSSSystem::UNKNOWN && group != GNSSSystem::GPS;
-}
-
-double groupDelayCorrectionMeters(const Observation& observation, const Ephemeris& eph) {
-    switch (observation.satellite.system) {
-        case GNSSSystem::GPS:
-        case GNSSSystem::QZSS:
-        case GNSSSystem::Galileo:
-            return eph.tgd * constants::SPEED_OF_LIGHT;
-        case GNSSSystem::BeiDou:
-            switch (observation.signal) {
-                case SignalType::BDS_B1I:
-                case SignalType::BDS_B1C:
-                    return eph.tgd * constants::SPEED_OF_LIGHT;
-                case SignalType::BDS_B2I:
-                case SignalType::BDS_B2A:
-                    return eph.tgd_secondary * constants::SPEED_OF_LIGHT;
-                default:
-                    return 0.0;
-            }
-        default:
-            return 0.0;
-    }
-}
-
-bool isHealthyForPositioning(const Observation& observation, const Ephemeris& eph) {
-    int sv_health = static_cast<int>(eph.health);
-    if (observation.satellite.system == GNSSSystem::QZSS) {
-        sv_health &= 0xFE;
-    }
-    return sv_health == 0;
-}
-
-Vector3d earthRotationCorrected(const Vector3d& satellite_position,
-                                const Vector3d& receiver_position) {
-    const double signal_travel_time =
-        (satellite_position - receiver_position).norm() / constants::SPEED_OF_LIGHT;
-    const double angle = constants::OMEGA_E * signal_travel_time;
-
-    Eigen::Matrix3d earth_rotation;
-    earth_rotation << std::cos(angle),  std::sin(angle), 0.0,
-                     -std::sin(angle),  std::cos(angle), 0.0,
-                      0.0,              0.0,             1.0;
-    return earth_rotation * satellite_position;
-}
-
-PositionSolution makeInvalidSolution(const GNSSTime& time) {
-    PositionSolution solution;
-    solution.time = time;
-    solution.status = SolutionStatus::NONE;
-    return solution;
-}
-
-Eigen::MatrixXd pseudoInverse(const Eigen::MatrixXd& matrix, double tolerance = 1e-12) {
-    if (matrix.size() == 0) {
-        return Eigen::MatrixXd();
-    }
-
-    Eigen::JacobiSVD<Eigen::MatrixXd> svd(
-        matrix,
-        Eigen::ComputeThinU | Eigen::ComputeThinV);
-    const auto& singular_values = svd.singularValues();
-    Eigen::VectorXd inverse_values = Eigen::VectorXd::Zero(singular_values.size());
-    const double threshold =
-        tolerance * std::max(matrix.rows(), matrix.cols()) *
-        (singular_values.size() > 0 ? singular_values.array().abs().maxCoeff() : 0.0);
-
-    for (int i = 0; i < singular_values.size(); ++i) {
-        if (singular_values(i) > threshold) {
-            inverse_values(i) = 1.0 / singular_values(i);
-        }
-    }
-
-    return svd.matrixV() * inverse_values.asDiagonal() * svd.matrixU().transpose();
-}
-
-double seedPositionDivergenceMeters(
-    const Vector3d& position_ecef,
-    const FGOProcessor::EpochSeed& seed) {
-    if (!position_ecef.allFinite() || !seed.position_ecef.allFinite() ||
-        seed.position_ecef.norm() <= 1e6) {
-        return std::numeric_limits<double>::quiet_NaN();
-    }
-    return (position_ecef - seed.position_ecef).norm();
-}
-
-bool applyFloatSeedPositionDivergenceGate(
-    PositionSolution& solution,
-    const FGOProcessor::FGOProblem& problem,
-    std::size_t epoch_index,
-    double max_divergence_m,
-    std::size_t& rejected_count) {
-    if (solution.status != SolutionStatus::FLOAT ||
-        !std::isfinite(max_divergence_m) || max_divergence_m <= 0.0) {
-        return false;
-    }
-    if (!solution.position_ecef.allFinite()) {
-        solution.status = SolutionStatus::NONE;
-        ++rejected_count;
-        return true;
-    }
-    if (epoch_index >= problem.epochs.size()) {
-        return false;
-    }
-
-    const double divergence_m =
-        seedPositionDivergenceMeters(solution.position_ecef,
-                                     problem.epochs[epoch_index]);
-    if (std::isfinite(divergence_m) && divergence_m > max_divergence_m) {
-        solution.status = SolutionStatus::NONE;
-        ++rejected_count;
-        return true;
-    }
-    return false;
-}
-
-bool applyFloatPositionJumpGate(
-    PositionSolution& solution,
-    const Vector3d& previous_output_position,
-    bool has_previous_output_position,
-    double max_jump_m,
-    bool& block_float_until_fixed,
-    std::size_t& rejected_count) {
-    if (!std::isfinite(max_jump_m) || max_jump_m <= 0.0) {
-        return false;
-    }
-    if (solution.status == SolutionStatus::FIXED) {
-        block_float_until_fixed = false;
-        return false;
-    }
-    if (solution.status != SolutionStatus::FLOAT) {
-        return false;
-    }
-
-    bool reject = block_float_until_fixed;
-    if (has_previous_output_position &&
-        solution.position_ecef.allFinite() &&
-        previous_output_position.allFinite()) {
-        const double jump_m =
-            (solution.position_ecef - previous_output_position).norm();
-        if (std::isfinite(jump_m) && jump_m > max_jump_m) {
-            reject = true;
-            block_float_until_fixed = true;
-        }
-    }
-    if (reject) {
-        solution.status = SolutionStatus::NONE;
-        ++rejected_count;
-        return true;
-    }
-    return false;
-}
-
-struct DoubleDifferencePrediction {
-    bool valid = false;
-    double geometry_m = 0.0;
-    Vector3d position_jacobian = Vector3d::Zero();
-};
-
-DoubleDifferencePrediction doubleDifferencePredictionAt(
-    const Vector3d& position,
-    const Vector3d& base_position,
-    const Vector3d& rover_satellite_position,
-    const Vector3d& rover_reference_position,
-    const Vector3d& base_satellite_position,
-    const Vector3d& base_reference_position) {
-    DoubleDifferencePrediction prediction;
-    const Vector3d rover_satellite_delta = rover_satellite_position - position;
-    const Vector3d rover_reference_delta = rover_reference_position - position;
-    const double rover_satellite_range = rover_satellite_delta.norm();
-    const double rover_reference_range = rover_reference_delta.norm();
-    const double base_satellite_range =
-        (base_satellite_position - base_position).norm();
-    const double base_reference_range =
-        (base_reference_position - base_position).norm();
-    if (rover_satellite_range <= 0.0 || rover_reference_range <= 0.0 ||
-        base_satellite_range <= 0.0 || base_reference_range <= 0.0) {
-        return prediction;
-    }
-
-    const Vector3d rover_satellite_los =
-        rover_satellite_delta / rover_satellite_range;
-    const Vector3d rover_reference_los =
-        rover_reference_delta / rover_reference_range;
-    prediction.valid = true;
-    prediction.geometry_m =
-        (rover_satellite_range - base_satellite_range) -
-        (rover_reference_range - base_reference_range);
-    prediction.position_jacobian = -rover_satellite_los + rover_reference_los;
-    return prediction;
-}
-
-struct PreparedCarrierObservation {
-    SatelliteId satellite;
-    SignalType signal = SignalType::GPS_L1CA;
-    Vector3d satellite_position_ecef = Vector3d::Zero();
-    double corrected_pseudorange_m = 0.0;
-    double corrected_carrier_m = 0.0;
-    double wavelength_m = 0.0;
-    double sigma_m = 0.01;
-    double elevation_rad = 0.0;
-    bool loss_of_lock = false;
-    bool has_carrier_phase = true;
-    bool has_doppler_residual = false;
-    double doppler_residual_mps = 0.0;
-    double doppler_sigma_mps = 0.2;
-    Vector3d los = Vector3d::Zero();
-    FGOProcessor::ObservationModelDebug model_debug;
-};
-
-struct ActiveCarrierSegment {
-    std::size_t ambiguity_index = 0;
-    std::size_t last_epoch_index = 0;
-};
-
-// --- Elevation-dependent DD sigma ("varerr") ---
-//
-// Port of the inuex35 reference's preprocess/prefit.py varerr_dd_sigma
-// (itself a port of RTKLIB-demo5's rtkpos.c:402 varerr()), gated by
-// FGOConfig::use_elevation_dependent_sigma (see that knob's doc comment in
-// fgo.hpp for the full formula, dt_s mapping, and composition-order
-// rationale). `is_pseudorange` selects the reference's `fact` term
-// (elevation_sigma_pseudorange_ratio for pseudorange, 1.0 for carrier);
-// `el_pair_rad` is the caller-computed max(min(el_ref, el_target), el_min);
-// `dt_s` is the rover epoch-to-epoch interval.
-double varerrDdSigma(bool is_pseudorange, double el_pair_rad, double dt_s,
-                     const FGOProcessor::FGOConfig& config) {
-    // No clamping on the ratio -- faithful to the reference's
-    // `fact = cfg.err_eratio_pr if code else 1.0` (no floor/ceiling there).
-    const double fact =
-        is_pseudorange ? config.elevation_sigma_pseudorange_ratio : 1.0;
-    const double a = fact * config.elevation_sigma_err_a_m;
-    const double b = fact * config.elevation_sigma_err_b_m;
-    const double d = constants::SPEED_OF_LIGHT * config.elevation_sigma_clock_stability * dt_s;
-    const double sinel = std::max(std::sin(el_pair_rad), 0.05);
-    const double var_sd = 2.0 * (a * a + b * b / (sinel * sinel)) + d * d;
-    return std::sqrt(2.0 * var_sd);
-}
-
-struct FixedAmbiguityConstraint {
-    std::size_t ambiguity_index = 0;
-    double fixed_ambiguity_m = 0.0;
-    int fixed_cycles = 0;
-    double residual_cycles = 0.0;
-    bool fixed_by_lambda = false;
-};
-
-using CarrierKey = std::pair<SatelliteId, SignalType>;
-
-// --- Code-Minus-Carrier (CMC) multipath screening state ---
-//
-// Port of inuex35's preprocess/slip_detect.py per-(satellite, signal) CMC
-// tracking (see FGOConfig::use_code_minus_carrier_screening for the full
-// semantics). Kept as function-local state in buildDoubleDifferenceProblem
-// (like active_dd_segments below) rather than a FGOProcessor member: one
-// build call processes one ordered epoch stream, and buildDoubleDifference-
-// Problem is const.
-struct CmcState {
-    bool has_prev = false;
-    double prev_cmc_m = 0.0;
-    bool has_baseline = false;
-    double baseline_m = 0.0;
-    int warmup_count = 0;
-};
-
-struct DoubleDifferenceAmbiguityKey {
-    SatelliteId satellite;
-    SatelliteId reference_satellite;
-    SignalType signal = SignalType::GPS_L1CA;
-    std::size_t satellite_ambiguity_index = 0;
-    std::size_t reference_ambiguity_index = 0;
-
-    bool operator<(const DoubleDifferenceAmbiguityKey& other) const {
-        return std::tie(satellite,
-                        reference_satellite,
-                        signal,
-                        satellite_ambiguity_index,
-                        reference_ambiguity_index) <
-               std::tie(other.satellite,
-                        other.reference_satellite,
-                        other.signal,
-                        other.satellite_ambiguity_index,
-                        other.reference_ambiguity_index);
-    }
-};
-
-struct DoubleDifferenceSegmentKey {
-    SatelliteId satellite;
-    SatelliteId reference_satellite;
-    SignalType signal = SignalType::GPS_L1CA;
-
-    bool operator<(const DoubleDifferenceSegmentKey& other) const {
-        return std::tie(satellite, reference_satellite, signal) <
-               std::tie(other.satellite, other.reference_satellite, other.signal);
-    }
-};
-
-struct SingleDifferenceReferenceKey {
-    std::size_t epoch_index = 0;
-    GNSSSystem system = GNSSSystem::UNKNOWN;
-    SignalType signal = SignalType::GPS_L1CA;
-
-    bool operator<(const SingleDifferenceReferenceKey& other) const {
-        return std::tie(epoch_index, system, signal) <
-               std::tie(other.epoch_index, other.system, other.signal);
-    }
-};
-
-struct SingleDifferenceDefaultReferenceKey {
-    GNSSSystem system = GNSSSystem::UNKNOWN;
-    SignalType signal = SignalType::GPS_L1CA;
-
-    bool operator<(const SingleDifferenceDefaultReferenceKey& other) const {
-        return std::tie(system, signal) <
-               std::tie(other.system, other.signal);
-    }
-};
-
-struct SingleDifferenceCarrierResidual {
-    std::size_t epoch_index = 0;
-    double residual_m = 0.0;
-    Vector3d los = Vector3d::Zero();
-};
-
-struct GeometryFreeSlipState {
-    GNSSTime time;
-    double geometry_free_m = 0.0;
-    std::size_t first_rover_ambiguity_index = 0;
-    std::size_t second_rover_ambiguity_index = 0;
-};
-
-struct GeometryFreePendingReset {
-    GNSSTime time;
-    std::size_t rover_ambiguity_index = 0;
-};
-
-std::map<CarrierKey, PreparedCarrierObservation> prepareCarrierObservationsForReceiver(
-    const ObservationData& epoch,
-    const NavigationData& nav,
-    const Vector3d& receiver_position,
-    const FGOProcessor::FGOConfig& config,
-    double min_snr_dbhz,
-    bool require_carrier_phase = true,
-    bool apply_elevation_mask = true,
-    const SecondaryCodeTable* secondary_code_table = nullptr) {
-    std::map<CarrierKey, PreparedCarrierObservation> carriers;
-    if (receiver_position.norm() <= 1e6) {
-        return carriers;
-    }
-
-    double receiver_lat = 0.0;
-    double receiver_lon = 0.0;
-    double receiver_height = 0.0;
-    ecef2geodetic(receiver_position, receiver_lat, receiver_lon, receiver_height);
-
-    const double min_elevation_rad = config.min_elevation_deg * M_PI / 180.0;
-    for (const auto& observation : epoch.observations) {
-        if (!isEligibleFgoSignal(observation.satellite,
-                                 observation.signal,
-                                 config.use_multi_constellation,
-                                 config.use_multi_frequency_double_difference)) {
-            continue;
-        }
-        if (!passesSecondaryCodeAlignment(observation, secondary_code_table)) {
-            continue;
-        }
-        const bool has_carrier_phase =
-            observation.has_carrier_phase && observation.carrier_phase != 0.0;
-        if (!observation.valid || !observation.has_pseudorange ||
-            observation.pseudorange <= 0.0 || observation.snr < min_snr_dbhz ||
-            (require_carrier_phase && !has_carrier_phase)) {
-            continue;
-        }
-
-        Vector3d satellite_position;
-        Vector3d satellite_velocity;
-        double satellite_clock_bias = 0.0;
-        double satellite_clock_drift = 0.0;
-
-        GNSSTime transmit_time =
-            epoch.time - observation.pseudorange / constants::SPEED_OF_LIGHT;
-        if (!nav.calculateSatelliteState(observation.satellite,
-                                         transmit_time,
-                                         satellite_position,
-                                         satellite_velocity,
-                                         satellite_clock_bias,
-                                         satellite_clock_drift)) {
-            continue;
-        }
-
-        transmit_time = transmit_time - satellite_clock_bias;
-        if (!nav.calculateSatelliteState(observation.satellite,
-                                         transmit_time,
-                                         satellite_position,
-                                         satellite_velocity,
-                                         satellite_clock_bias,
-                                         satellite_clock_drift)) {
-            continue;
-        }
-
-        const Ephemeris* eph = nav.getEphemeris(observation.satellite, transmit_time);
-        if (!eph || !isHealthyForPositioning(observation, *eph)) {
-            continue;
-        }
-
-        const Vector3d corrected_satellite_position =
-            earthRotationCorrected(satellite_position, receiver_position);
-        const auto geometry =
-            nav.calculateGeometry(receiver_position, corrected_satellite_position);
-        if (apply_elevation_mask && geometry.elevation < min_elevation_rad) {
-            continue;
-        }
-
-        double ionosphere_delay = 0.0;
-        if (config.use_ionosphere_model && nav.ionosphere_model.valid) {
-            ionosphere_delay = models::ionoDelayKlobuchar(
-                receiver_lat,
-                receiver_lon,
-                geometry.azimuth,
-                geometry.elevation,
-                epoch.time.tow,
-                nav.ionosphere_model.alpha,
-                nav.ionosphere_model.beta);
-
-            const double frequency_hz = signalFrequencyHz(observation.signal, eph);
-            if (frequency_hz > 0.0) {
-                const double scale = constants::GPS_L1_FREQ / frequency_hz;
-                ionosphere_delay *= scale * scale;
-            }
-        }
-
-        double troposphere_delay = 0.0;
-        if (config.use_troposphere_model) {
-            troposphere_delay =
-                models::tropDelaySaastamoinen(receiver_position, geometry.elevation);
-        }
-
-        double wavelength = signalWavelengthMeters(observation);
-        if (wavelength <= 0.0) {
-            wavelength = signalWavelengthMeters(observation.signal, eph);
-        }
-        const bool usable_carrier = has_carrier_phase && wavelength > 0.0;
-        if (require_carrier_phase && !usable_carrier) {
-            continue;
-        }
-
-        const double satellite_clock_m =
-            satellite_clock_bias * constants::SPEED_OF_LIGHT;
-        const double group_delay_m = groupDelayCorrectionMeters(observation, *eph);
-        const double corrected_pseudorange =
-            observation.pseudorange +
-            satellite_clock_m -
-            ionosphere_delay -
-            troposphere_delay -
-            group_delay_m;
-        const double corrected_carrier =
-            usable_carrier
-                ? observation.carrier_phase * wavelength +
-                      satellite_clock_m -
-                      troposphere_delay +
-                      ionosphere_delay
-                : 0.0;
-
-        FGOProcessor::ObservationModelDebug model_debug;
-        model_debug.raw_pseudorange_m = observation.pseudorange;
-        model_debug.raw_carrier_m =
-            usable_carrier ? observation.carrier_phase * wavelength : 0.0;
-        model_debug.satellite_clock_m = satellite_clock_m;
-        model_debug.ionosphere_delay_m = ionosphere_delay;
-        model_debug.troposphere_delay_m = troposphere_delay;
-        model_debug.group_delay_m = group_delay_m;
-        model_debug.corrected_pseudorange_m = corrected_pseudorange;
-        model_debug.corrected_carrier_m = corrected_carrier;
-        const Vector3d corrected_delta =
-            corrected_satellite_position - receiver_position;
-        const double corrected_range = corrected_delta.norm();
-        if (corrected_range <= 0.0) {
-            continue;
-        }
-        model_debug.geometric_range_m = corrected_range;
-        model_debug.elevation_rad = geometry.elevation;
-        model_debug.azimuth_rad = geometry.azimuth;
-        model_debug.snr_dbhz = observation.snr;
-
-        const double sin_el = std::max(0.1, std::sin(geometry.elevation));
-        PreparedCarrierObservation carrier;
-        carrier.satellite = observation.satellite;
-        carrier.signal = observation.signal;
-        carrier.satellite_position_ecef = corrected_satellite_position;
-        carrier.corrected_pseudorange_m = corrected_pseudorange;
-        carrier.corrected_carrier_m = corrected_carrier;
-        carrier.wavelength_m = wavelength;
-        carrier.sigma_m = std::max(1e-4, config.carrier_phase_sigma_m / sin_el);
-        carrier.elevation_rad = geometry.elevation;
-        carrier.loss_of_lock =
-            observation.loss_of_lock || ((observation.lli & 0x01U) != 0);
-        carrier.has_carrier_phase = usable_carrier;
-        carrier.los = -corrected_delta / corrected_range;
-        carrier.doppler_sigma_mps =
-            std::max(1e-4, config.single_difference_doppler_sigma_mps /
-                               std::sqrt(sin_el));
-        if (observation.has_doppler && wavelength > 0.0) {
-            const Vector3d doppler_delta = satellite_position - receiver_position;
-            const double doppler_range = doppler_delta.norm();
-            if (doppler_range > 0.0) {
-                const Vector3d ex = doppler_delta / doppler_range;
-                const double sagnac_rate =
-                    constants::OMEGA_E / constants::SPEED_OF_LIGHT *
-                    (satellite_velocity(1) * receiver_position(0) -
-                     satellite_velocity(0) * receiver_position(1));
-                const double modeled_range_rate =
-                    satellite_velocity.dot(ex) + sagnac_rate;
-                const double satellite_clock_drift_mps =
-                    satellite_clock_drift * constants::SPEED_OF_LIGHT;
-                const double measured_range_rate =
-                    -observation.doppler * wavelength;
-                carrier.doppler_residual_mps =
-                    measured_range_rate -
-                    (modeled_range_rate - satellite_clock_drift_mps);
-                carrier.has_doppler_residual =
-                    std::isfinite(carrier.doppler_residual_mps);
-                model_debug.has_doppler_residual =
-                    carrier.has_doppler_residual;
-                model_debug.doppler_residual_mps =
-                    carrier.doppler_residual_mps;
-            }
-        }
-        carrier.model_debug = model_debug;
-        carriers[{observation.satellite, observation.signal}] = carrier;
-    }
-
-    return carriers;
-}
-
-Observation interpolateObservation(const Observation& lower,
-                                   const Observation& upper,
-                                   double alpha) {
-    Observation interpolated = lower;
-    interpolated.valid = lower.valid && upper.valid;
-    interpolated.has_pseudorange =
-        lower.has_pseudorange && upper.has_pseudorange;
-    interpolated.has_carrier_phase =
-        lower.has_carrier_phase && upper.has_carrier_phase;
-    interpolated.has_doppler = lower.has_doppler && upper.has_doppler;
-    if (interpolated.has_pseudorange) {
-        interpolated.pseudorange =
-            lower.pseudorange + alpha * (upper.pseudorange - lower.pseudorange);
-    }
-    if (interpolated.has_carrier_phase) {
-        interpolated.carrier_phase =
-            lower.carrier_phase +
-            alpha * (upper.carrier_phase - lower.carrier_phase);
-    }
-    if (interpolated.has_doppler) {
-        interpolated.doppler =
-            lower.doppler + alpha * (upper.doppler - lower.doppler);
-    }
-    interpolated.snr = std::min(lower.snr, upper.snr);
-    interpolated.signal_strength =
-        std::min(lower.signal_strength, upper.signal_strength);
-    interpolated.lli = lower.lli | upper.lli;
-    interpolated.loss_of_lock =
-        lower.loss_of_lock || upper.loss_of_lock ||
-        ((interpolated.lli & 0x01U) != 0);
-    interpolated.has_glonass_frequency_channel =
-        lower.has_glonass_frequency_channel &&
-        upper.has_glonass_frequency_channel &&
-        lower.glonass_frequency_channel == upper.glonass_frequency_channel;
-    // MF hygiene: never blend two different observation codes of the same
-    // band across the interpolation boundary (e.g. BDS B2I C7D at t0 and C7I
-    // at t1) -- the blend would carry an inter-code bias under a single code
-    // label. Secondary bands only, so the single-frequency (primary band)
-    // path is bit-identical.
-    if (signal_policy::isSecondarySignal(lower.satellite.system, lower.signal) &&
-        (lower.pseudorange_observation_type !=
-             upper.pseudorange_observation_type ||
-         lower.carrier_phase_observation_type !=
-             upper.carrier_phase_observation_type)) {
-        interpolated.valid = false;
-    }
-    return interpolated;
-}
-
-bool findMatchedBaseEpoch(const std::vector<ObservationData>& base_epochs,
-                          const GNSSTime& rover_time,
-                          std::size_t& cursor,
-                          double exact_tolerance_s,
-                          double interpolation_max_gap_s,
-                          ObservationData& matched_epoch,
-                          bool& interpolated) {
-    if (base_epochs.empty()) {
-        return false;
-    }
-
-    while (cursor + 1 < base_epochs.size() &&
-           base_epochs[cursor + 1].time <= rover_time) {
-        ++cursor;
-    }
-
-    const ObservationData* exact_match = nullptr;
-    double best_abs_dt = std::numeric_limits<double>::infinity();
-    auto consider_exact = [&](std::size_t index) {
-        if (index >= base_epochs.size()) {
-            return;
-        }
-        const double dt = std::abs(base_epochs[index].time - rover_time);
-        if (dt <= exact_tolerance_s && dt < best_abs_dt) {
-            best_abs_dt = dt;
-            exact_match = &base_epochs[index];
-        }
-    };
-    consider_exact(cursor);
-    consider_exact(cursor + 1);
-    if (cursor > 0) {
-        consider_exact(cursor - 1);
-    }
-
-    if (exact_match != nullptr) {
-        matched_epoch = *exact_match;
-        interpolated = false;
-        return true;
-    }
-
-    if (interpolation_max_gap_s <= 0.0) {
-        return false;
-    }
-
-    std::size_t lower_index = cursor;
-    if (base_epochs[lower_index].time > rover_time) {
-        if (lower_index == 0) {
-            return false;
-        }
-        --lower_index;
-    }
-    const std::size_t upper_index = lower_index + 1;
-    if (upper_index >= base_epochs.size()) {
-        return false;
-    }
-
-    const ObservationData& lower_epoch = base_epochs[lower_index];
-    const ObservationData& upper_epoch = base_epochs[upper_index];
-    const double total_dt = upper_epoch.time - lower_epoch.time;
-    const double lower_dt = rover_time - lower_epoch.time;
-    if (total_dt <= 0.0 || total_dt > interpolation_max_gap_s ||
-        lower_dt < 0.0 || lower_dt > total_dt) {
-        return false;
-    }
-
-    const double alpha = lower_dt / total_dt;
-    std::map<CarrierKey, const Observation*> upper_observations;
-    for (const auto& observation : upper_epoch.observations) {
-        upper_observations[{observation.satellite, observation.signal}] =
-            &observation;
-    }
-
-    ObservationData candidate(rover_time);
-    candidate.receiver_position = lower_epoch.receiver_position;
-    if (candidate.receiver_position.norm() <= 1e6) {
-        candidate.receiver_position = upper_epoch.receiver_position;
-    }
-    candidate.receiver_clock_bias =
-        lower_epoch.receiver_clock_bias +
-        alpha * (upper_epoch.receiver_clock_bias -
-                 lower_epoch.receiver_clock_bias);
-
-    for (const auto& lower_observation : lower_epoch.observations) {
-        const CarrierKey key{lower_observation.satellite, lower_observation.signal};
-        const auto upper_it = upper_observations.find(key);
-        if (upper_it == upper_observations.end()) {
-            continue;
-        }
-        const Observation& upper_observation = *upper_it->second;
-        if (!lower_observation.valid || !upper_observation.valid ||
-            !lower_observation.has_pseudorange ||
-            !upper_observation.has_pseudorange ||
-            !lower_observation.has_carrier_phase ||
-            !upper_observation.has_carrier_phase) {
-            continue;
-        }
-        candidate.addObservation(
-            interpolateObservation(lower_observation, upper_observation, alpha));
-    }
-
-    if (candidate.observations.empty()) {
-        return false;
-    }
-
-    matched_epoch = std::move(candidate);
-    interpolated = true;
-    return true;
-}
-
-}  // namespace
-
-FGOProcessor::FGOProblem FGOProcessor::buildPseudorangeProblem(
-    const std::vector<ObservationData>& input_epochs,
-    const NavigationData& nav) const {
-    FGOProblem problem;
-    problem.diagnostics.input_epochs = input_epochs.size();
-    if (input_epochs.empty()) {
-        return problem;
-    }
-
-    // MF hygiene: one observation code per (system, secondary band) for this
-    // receiver (cssrlib/RTKLIB-style). Only meaningful when secondary bands
-    // are ingested at all.
-    SecondaryCodeTable rover_secondary_codes;
-    const SecondaryCodeTable* rover_secondary_codes_ptr = nullptr;
-    if (config_.use_multi_frequency_double_difference &&
-        config_.use_double_difference_secondary_code_alignment &&
-        config_.use_multi_constellation) {
-        rover_secondary_codes = buildSecondaryCodeTable(input_epochs);
-        rover_secondary_codes_ptr = &rover_secondary_codes;
-    }
-
-    ProcessorConfig spp_processor_config;
-    spp_processor_config.mode = PositioningMode::SPP;
-    spp_processor_config.elevation_mask = config_.min_elevation_deg;
-    spp_processor_config.snr_mask = config_.min_snr_dbhz;
-    spp_processor_config.use_ionosphere_model = config_.use_ionosphere_model;
-    spp_processor_config.use_troposphere_model = config_.use_troposphere_model;
-
-    SPPProcessor::SPPConfig spp_config;
-    spp_config.apply_atmospheric_corrections =
-        config_.use_ionosphere_model || config_.use_troposphere_model;
-    spp_config.use_multi_constellation = config_.use_multi_constellation;
-
-    SPPProcessor spp_processor(spp_config);
-    spp_processor.initialize(spp_processor_config);
-
-    const double min_elevation_rad = config_.min_elevation_deg * M_PI / 180.0;
-    const double dd_reference_min_snr_dbhz =
-        config_.double_difference_reference_min_snr_dbhz >= 0.0
-            ? config_.double_difference_reference_min_snr_dbhz
-            : config_.min_snr_dbhz;
-    std::vector<std::map<std::pair<SatelliteId, SignalType>, PreparedCarrierObservation>>
-        carrier_by_problem_epoch;
-    std::vector<std::map<std::pair<SatelliteId, SignalType>, PreparedCarrierObservation>>
-        dd_pseudorange_by_problem_epoch;
-    std::vector<std::map<std::pair<SatelliteId, SignalType>, PreparedCarrierObservation>>
-        dd_reference_carrier_by_problem_epoch;
-    std::map<SatelliteId, double> previous_gps_pseudorange_by_satellite;
-
-    // Seed continuity across a brief SPP outage (see FGOConfig::use_spp_seed):
-    // a tunnel/underpass/urban-canyon gap can drop the rover to near-zero
-    // tracked satellites for several seconds. The RINEX header's static
-    // approximate position can be kilometres from the true kinematic position
-    // at that point in the trajectory; seeding elevation/geometry from it
-    // spuriously fails the elevation mask for whatever few satellites ARE
-    // still tracked during recovery, rejecting far more marginal epochs than
-    // the min_satellites_per_epoch gate actually requires. Coast on the last
-    // valid SPP fix instead -- at the rover's epoch-to-epoch cadence this
-    // stays far closer to truth through a brief outage than the static
-    // fallback, and degrades no worse than the static fallback once the
-    // outage is long enough that both are stale.
-    Vector3d last_valid_seed_position_ecef = Vector3d::Zero();
-    double last_valid_seed_clock_bias_m = 0.0;
-    bool have_last_valid_seed = false;
-
-    for (const auto& epoch : input_epochs) {
-        EpochSeed seed;
-        seed.time = epoch.time;
-
-        PositionSolution spp_solution = makeInvalidSolution(epoch.time);
-        if (config_.use_spp_seed) {
-            spp_solution = spp_processor.processEpoch(epoch, nav);
-        }
-
-        if (spp_solution.isValid()) {
-            seed.position_ecef = spp_solution.position_ecef;
-            seed.receiver_clock_bias_m = spp_solution.receiver_clock_bias;
-            seed.fresh_spp_solution = true;
-            last_valid_seed_position_ecef = seed.position_ecef;
-            last_valid_seed_clock_bias_m = seed.receiver_clock_bias_m;
-            have_last_valid_seed = true;
-        } else if (have_last_valid_seed) {
-            seed.position_ecef = last_valid_seed_position_ecef;
-            seed.receiver_clock_bias_m = last_valid_seed_clock_bias_m;
-        } else if (epoch.receiver_position.norm() > 1e6) {
-            seed.position_ecef = epoch.receiver_position;
-            seed.receiver_clock_bias_m = epoch.receiver_clock_bias * constants::SPEED_OF_LIGHT;
-        } else {
-            ++problem.diagnostics.skipped_epochs_without_seed;
-            continue;
-        }
-
-        std::vector<PseudorangeFactor> epoch_factors;
-        epoch_factors.reserve(epoch.observations.size());
-        std::map<std::pair<SatelliteId, SignalType>, PreparedCarrierObservation> epoch_carriers;
-        std::map<std::pair<SatelliteId, SignalType>, PreparedCarrierObservation>
-            epoch_dd_pseudorange_observations;
-        std::map<std::pair<SatelliteId, SignalType>, PreparedCarrierObservation>
-            epoch_dd_reference_carriers;
-        std::map<SatelliteId, double> gps_pseudorange_by_satellite;
-
-        double receiver_lat = 0.0;
-        double receiver_lon = 0.0;
-        double receiver_height = 0.0;
-        ecef2geodetic(seed.position_ecef, receiver_lat, receiver_lon, receiver_height);
-
-        for (const auto& observation : epoch.observations) {
-            if (!isEligibleFgoSignal(observation.satellite,
-                                     observation.signal,
-                                     config_.use_multi_constellation,
-                                     config_.use_multi_frequency_double_difference)) {
-                continue;
-            }
-            if (!passesSecondaryCodeAlignment(observation,
-                                              rover_secondary_codes_ptr)) {
-                continue;
-            }
-            if (!observation.valid || !observation.has_pseudorange ||
-                observation.pseudorange <= 0.0) {
-                continue;
-            }
-            const bool passes_snr_mask = observation.snr >= config_.min_snr_dbhz;
-            const bool passes_dd_reference_snr_mask =
-                dd_reference_min_snr_dbhz <= 0.0 ||
-                observation.snr >= dd_reference_min_snr_dbhz;
-
-            Vector3d satellite_position;
-            Vector3d satellite_velocity;
-            double satellite_clock_bias = 0.0;
-            double satellite_clock_drift = 0.0;
-
-            GNSSTime transmit_time =
-                epoch.time - observation.pseudorange / constants::SPEED_OF_LIGHT;
-            if (!nav.calculateSatelliteState(observation.satellite,
-                                             transmit_time,
-                                             satellite_position,
-                                             satellite_velocity,
-                                             satellite_clock_bias,
-                                             satellite_clock_drift)) {
-                continue;
-            }
-
-            transmit_time = transmit_time - satellite_clock_bias;
-            if (!nav.calculateSatelliteState(observation.satellite,
-                                             transmit_time,
-                                             satellite_position,
-                                             satellite_velocity,
-                                             satellite_clock_bias,
-                                             satellite_clock_drift)) {
-                continue;
-            }
-
-            const Ephemeris* eph = nav.getEphemeris(observation.satellite, transmit_time);
-            if (!eph || !isHealthyForPositioning(observation, *eph)) {
-                continue;
-            }
-
-            const Vector3d corrected_satellite_position =
-                earthRotationCorrected(satellite_position, seed.position_ecef);
-            const auto geometry =
-                nav.calculateGeometry(seed.position_ecef, corrected_satellite_position);
-            const bool passes_elevation_mask =
-                geometry.elevation >= min_elevation_rad;
-
-            double ionosphere_delay = 0.0;
-            if (config_.use_ionosphere_model && nav.ionosphere_model.valid) {
-                ionosphere_delay = models::ionoDelayKlobuchar(
-                    receiver_lat,
-                    receiver_lon,
-                    geometry.azimuth,
-                    geometry.elevation,
-                    epoch.time.tow,
-                    nav.ionosphere_model.alpha,
-                    nav.ionosphere_model.beta);
-
-                const double frequency_hz = signalFrequencyHz(observation.signal, eph);
-                if (frequency_hz > 0.0) {
-                    const double scale = constants::GPS_L1_FREQ / frequency_hz;
-                    ionosphere_delay *= scale * scale;
-                }
-            }
-
-            double troposphere_delay = 0.0;
-            if (config_.use_troposphere_model) {
-                troposphere_delay =
-                    models::tropDelaySaastamoinen(seed.position_ecef, geometry.elevation);
-            }
-
-            const double satellite_clock_m =
-                satellite_clock_bias * constants::SPEED_OF_LIGHT;
-            const double group_delay_m =
-                groupDelayCorrectionMeters(observation, *eph);
-            const double corrected_pseudorange =
-                observation.pseudorange +
-                satellite_clock_m -
-                ionosphere_delay -
-                troposphere_delay -
-                group_delay_m;
-
-            const double sin_el = std::max(0.1, std::sin(geometry.elevation));
-            const double pseudorange_elevation_scale = std::pow(
-                sin_el,
-                std::max(0.0, config_.pseudorange_elevation_sigma_power));
-            if (passes_snr_mask && passes_elevation_mask) {
-                PseudorangeFactor factor;
-                factor.satellite = observation.satellite;
-                factor.clock_group = clockBiasGroup(observation.satellite.system);
-                factor.satellite_position_ecef = corrected_satellite_position;
-                factor.corrected_pseudorange_m = corrected_pseudorange;
-                factor.sigma_m =
-                    std::max(1e-3,
-                             config_.pseudorange_sigma_m /
-                                 std::max(1e-6, pseudorange_elevation_scale));
-                factor.elevation_rad = geometry.elevation;
-                epoch_factors.push_back(factor);
-                if (observation.satellite.system == GNSSSystem::GPS) {
-                    gps_pseudorange_by_satellite[observation.satellite] =
-                        observation.pseudorange;
-                }
-            }
-
-            const bool carrier_loss_of_lock =
-                observation.loss_of_lock || ((observation.lli & 0x01U) != 0);
-            double wavelength = signalWavelengthMeters(observation);
-            const bool has_carrier_phase =
-                observation.has_carrier_phase && observation.carrier_phase != 0.0;
-            if (wavelength <= 0.0) {
-                wavelength = signalWavelengthMeters(observation.signal, eph);
-            }
-            const bool usable_carrier = has_carrier_phase && wavelength > 0.0;
-            const double corrected_carrier =
-                usable_carrier
-                    ? observation.carrier_phase * wavelength +
-                          satellite_clock_m -
-                          troposphere_delay +
-                          ionosphere_delay
-                    : 0.0;
-
-            FGOProcessor::ObservationModelDebug model_debug;
-            model_debug.raw_pseudorange_m = observation.pseudorange;
-            model_debug.raw_carrier_m =
-                usable_carrier ? observation.carrier_phase * wavelength : 0.0;
-            model_debug.satellite_clock_m = satellite_clock_m;
-            model_debug.ionosphere_delay_m = ionosphere_delay;
-            model_debug.troposphere_delay_m = troposphere_delay;
-            model_debug.group_delay_m = group_delay_m;
-            model_debug.corrected_pseudorange_m = corrected_pseudorange;
-            model_debug.corrected_carrier_m = corrected_carrier;
-            const Vector3d corrected_delta =
-                corrected_satellite_position - seed.position_ecef;
-            const double corrected_range = corrected_delta.norm();
-            if (corrected_range <= 0.0) {
-                continue;
-            }
-            model_debug.geometric_range_m = corrected_range;
-            model_debug.elevation_rad = geometry.elevation;
-            model_debug.azimuth_rad = geometry.azimuth;
-            model_debug.snr_dbhz = observation.snr;
-
-            PreparedCarrierObservation carrier;
-            carrier.satellite = observation.satellite;
-            carrier.signal = observation.signal;
-            carrier.satellite_position_ecef = corrected_satellite_position;
-            carrier.corrected_pseudorange_m = corrected_pseudorange;
-            carrier.corrected_carrier_m = corrected_carrier;
-            carrier.wavelength_m = wavelength;
-            carrier.sigma_m =
-                std::max(1e-4, config_.carrier_phase_sigma_m / sin_el);
-            carrier.elevation_rad = geometry.elevation;
-            carrier.loss_of_lock = carrier_loss_of_lock;
-            carrier.has_carrier_phase = usable_carrier;
-            carrier.los = -corrected_delta / corrected_range;
-            carrier.doppler_sigma_mps =
-                std::max(1e-4,
-                         config_.single_difference_doppler_sigma_mps /
-                             std::sqrt(sin_el));
-            if (observation.has_doppler && wavelength > 0.0) {
-                const Vector3d doppler_delta =
-                    satellite_position - seed.position_ecef;
-                const double doppler_range = doppler_delta.norm();
-                if (doppler_range > 0.0) {
-                    const Vector3d ex = doppler_delta / doppler_range;
-                    const double sagnac_rate =
-                        constants::OMEGA_E / constants::SPEED_OF_LIGHT *
-                        (satellite_velocity(1) * seed.position_ecef(0) -
-                         satellite_velocity(0) * seed.position_ecef(1));
-                    const double modeled_range_rate =
-                        satellite_velocity.dot(ex) + sagnac_rate;
-                    const double satellite_clock_drift_mps =
-                        satellite_clock_drift * constants::SPEED_OF_LIGHT;
-                    const double measured_range_rate =
-                        -observation.doppler * wavelength;
-                    carrier.doppler_residual_mps =
-                        measured_range_rate -
-                        (modeled_range_rate - satellite_clock_drift_mps);
-                    carrier.has_doppler_residual =
-                        std::isfinite(carrier.doppler_residual_mps);
-                    model_debug.has_doppler_residual =
-                        carrier.has_doppler_residual;
-                    model_debug.doppler_residual_mps =
-                        carrier.doppler_residual_mps;
-                }
-            }
-            carrier.model_debug = model_debug;
-            if (passes_dd_reference_snr_mask) {
-                epoch_dd_reference_carriers[
-                    {observation.satellite, observation.signal}] = carrier;
-            }
-            if (passes_snr_mask && passes_elevation_mask) {
-                epoch_dd_pseudorange_observations[
-                    {observation.satellite, observation.signal}] = carrier;
-            }
-            if (usable_carrier && passes_snr_mask && passes_elevation_mask) {
-                if (!(config_.reject_rover_carrier_loss_of_lock &&
-                      carrier_loss_of_lock)) {
-                    epoch_carriers[{observation.satellite, observation.signal}] =
-                        carrier;
-                }
-            }
-        }
-
-        const std::size_t usable_epoch_measurements =
-            config_.use_pseudorange_factors ? epoch_factors.size()
-                                            : epoch_carriers.size();
-        if (static_cast<int>(usable_epoch_measurements) <
-            config_.min_satellites_per_epoch) {
-            continue;
-        }
-
-        const std::size_t epoch_index = problem.epochs.size();
-        problem.epochs.push_back(seed);
-        bool clock_jump = false;
-        double gps_pseudorange_delta_sum = 0.0;
-        std::size_t gps_pseudorange_delta_count = 0;
-        for (const auto& [satellite, pseudorange] : gps_pseudorange_by_satellite) {
-            const auto previous_it =
-                previous_gps_pseudorange_by_satellite.find(satellite);
-            if (previous_it == previous_gps_pseudorange_by_satellite.end()) {
-                continue;
-            }
-            gps_pseudorange_delta_sum += pseudorange - previous_it->second;
-            ++gps_pseudorange_delta_count;
-        }
-        if (gps_pseudorange_delta_count > 0) {
-            const double mean_delta =
-                gps_pseudorange_delta_sum /
-                static_cast<double>(gps_pseudorange_delta_count);
-            clock_jump = mean_delta > 1e5;
-            problem.gps_common_pseudorange_delta_m.push_back(mean_delta);
-        } else {
-            problem.gps_common_pseudorange_delta_m.push_back(0.0);
-        }
-        problem.gps_common_pseudorange_delta_satellites.push_back(
-            static_cast<int>(gps_pseudorange_delta_count));
-        problem.clock_jumps.push_back(clock_jump);
-        previous_gps_pseudorange_by_satellite =
-            std::move(gps_pseudorange_by_satellite);
-        problem.diagnostics.seeded_epochs = problem.epochs.size();
-        carrier_by_problem_epoch.push_back(std::move(epoch_carriers));
-        dd_pseudorange_by_problem_epoch.push_back(
-            std::move(epoch_dd_pseudorange_observations));
-        dd_reference_carrier_by_problem_epoch.push_back(
-            std::move(epoch_dd_reference_carriers));
-        if (config_.use_pseudorange_factors) {
-            for (auto& factor : epoch_factors) {
-                factor.epoch_index = epoch_index;
-                problem.pseudorange_factors.push_back(factor);
-            }
-        }
-    }
-
-    if ((config_.use_carrier_phase_factors || config_.use_double_difference_factors) &&
-        !carrier_by_problem_epoch.empty()) {
-        using CarrierKey = std::pair<SatelliteId, SignalType>;
-        std::map<CarrierKey, ActiveCarrierSegment> active_segments;
-        std::map<CarrierKey, std::size_t> next_segment_indices;
-        const double max_gap = std::max(0.0, config_.max_tdcp_gap_s);
-
-        for (std::size_t epoch_index = 0; epoch_index < problem.epochs.size(); ++epoch_index) {
-            const auto& current_carriers = carrier_by_problem_epoch[epoch_index];
-            for (const auto& [key, carrier] : current_carriers) {
-                auto active_it = active_segments.find(key);
-                bool start_new_segment = active_it == active_segments.end();
-                if (active_it != active_segments.end()) {
-                    const double dt = problem.epochs[epoch_index].time -
-                                      problem.epochs[active_it->second.last_epoch_index].time;
-                    if (dt <= 0.0 || (max_gap > 0.0 && dt > max_gap) ||
-                        carrier.loss_of_lock) {
-                        start_new_segment = true;
-                    }
-                }
-
-                if (start_new_segment) {
-                    const std::size_t segment_index = next_segment_indices[key]++;
-                    std::size_t carrier_segment_index = segment_index;
-                    if (config_.use_carrier_phase_factors) {
-                        AmbiguityState ambiguity;
-                        ambiguity.satellite = carrier.satellite;
-                        ambiguity.signal = carrier.signal;
-                        ambiguity.segment_index = segment_index;
-                        ambiguity.wavelength_m = carrier.wavelength_m;
-                        ambiguity.initial_ambiguity_m =
-                            carrier.corrected_carrier_m -
-                            carrier.corrected_pseudorange_m;
-                        carrier_segment_index = problem.ambiguity_states.size();
-                        problem.ambiguity_states.push_back(ambiguity);
-                    }
-                    active_it = active_segments
-                                    .insert_or_assign(
-                                        key,
-                                        ActiveCarrierSegment{carrier_segment_index,
-                                                             epoch_index})
-                                    .first;
-                } else {
-                    active_it->second.last_epoch_index = epoch_index;
-                }
-
-                CarrierPhaseFactor carrier_observation;
-                carrier_observation.epoch_index = epoch_index;
-                carrier_observation.ambiguity_index = active_it->second.ambiguity_index;
-                carrier_observation.satellite = carrier.satellite;
-                carrier_observation.clock_group =
-                    clockBiasGroup(carrier.satellite.system);
-                carrier_observation.signal = carrier.signal;
-                carrier_observation.satellite_position_ecef =
-                    carrier.satellite_position_ecef;
-                carrier_observation.corrected_pseudorange_m =
-                    carrier.corrected_pseudorange_m;
-                carrier_observation.corrected_carrier_m =
-                    carrier.corrected_carrier_m;
-                carrier_observation.wavelength_m = carrier.wavelength_m;
-                carrier_observation.sigma_m = carrier.sigma_m;
-                carrier_observation.elevation_rad = carrier.elevation_rad;
-                carrier_observation.has_carrier_phase = carrier.has_carrier_phase;
-                carrier_observation.loss_of_lock = carrier.loss_of_lock;
-                carrier_observation.has_doppler_residual =
-                    carrier.has_doppler_residual;
-                carrier_observation.doppler_residual_mps =
-                    carrier.doppler_residual_mps;
-                carrier_observation.doppler_sigma_mps =
-                    carrier.doppler_sigma_mps;
-                carrier_observation.los = carrier.los;
-                carrier_observation.model_debug = carrier.model_debug;
-                problem.carrier_observations.push_back(carrier_observation);
-                if (config_.use_carrier_phase_factors) {
-                    problem.carrier_phase_factors.push_back(carrier_observation);
-                }
-            }
-        }
-    }
-
-    if (config_.use_double_difference_factors) {
-        for (std::size_t epoch_index = 0;
-             epoch_index < dd_pseudorange_by_problem_epoch.size();
-             ++epoch_index) {
-            for (const auto& [key, carrier] :
-                 dd_pseudorange_by_problem_epoch[epoch_index]) {
-                CarrierPhaseFactor pseudorange_observation;
-                pseudorange_observation.epoch_index = epoch_index;
-                pseudorange_observation.satellite = carrier.satellite;
-                pseudorange_observation.clock_group =
-                    clockBiasGroup(carrier.satellite.system);
-                pseudorange_observation.signal = carrier.signal;
-                pseudorange_observation.satellite_position_ecef =
-                    carrier.satellite_position_ecef;
-                pseudorange_observation.corrected_pseudorange_m =
-                    carrier.corrected_pseudorange_m;
-                pseudorange_observation.corrected_carrier_m =
-                    carrier.corrected_carrier_m;
-                pseudorange_observation.wavelength_m = carrier.wavelength_m;
-                pseudorange_observation.sigma_m = carrier.sigma_m;
-                pseudorange_observation.elevation_rad = carrier.elevation_rad;
-                pseudorange_observation.has_carrier_phase =
-                    carrier.has_carrier_phase;
-                pseudorange_observation.loss_of_lock = carrier.loss_of_lock;
-                pseudorange_observation.has_doppler_residual =
-                    carrier.has_doppler_residual;
-                pseudorange_observation.doppler_residual_mps =
-                    carrier.doppler_residual_mps;
-                pseudorange_observation.doppler_sigma_mps =
-                    carrier.doppler_sigma_mps;
-                pseudorange_observation.los = carrier.los;
-                pseudorange_observation.model_debug = carrier.model_debug;
-                problem.double_difference_pseudorange_observations.push_back(
-                    pseudorange_observation);
-            }
-        }
-    }
-
-    if (config_.use_tdcp_factors && problem.epochs.size() >= 2) {
-        const double sigma = std::max(1e-4, config_.tdcp_sigma_m);
-        const double max_gap = std::max(0.0, config_.max_tdcp_gap_s);
-        for (std::size_t epoch_index = 1; epoch_index < problem.epochs.size(); ++epoch_index) {
-            const double dt = problem.epochs[epoch_index].time -
-                              problem.epochs[epoch_index - 1].time;
-            if (dt <= 0.0 || (max_gap > 0.0 && dt > max_gap)) {
-                problem.diagnostics.tdcp_rejected_gap +=
-                    carrier_by_problem_epoch[epoch_index].size();
-                continue;
-            }
-
-            const auto& previous_carriers = carrier_by_problem_epoch[epoch_index - 1];
-            const auto& current_carriers = carrier_by_problem_epoch[epoch_index];
-            for (const auto& [key, current] : current_carriers) {
-                const auto previous_it = previous_carriers.find(key);
-                if (previous_it == previous_carriers.end()) {
-                    ++problem.diagnostics.tdcp_rejected_missing_previous;
-                    continue;
-                }
-                const auto& previous = previous_it->second;
-                ++problem.diagnostics.tdcp_candidate_pairs;
-                if (config_.reject_tdcp_loss_of_lock &&
-                    (previous.loss_of_lock || current.loss_of_lock)) {
-                    ++problem.diagnostics.tdcp_rejected_loss_of_lock;
-                    continue;
-                }
-
-                const double delta_carrier_m =
-                    current.corrected_carrier_m - previous.corrected_carrier_m;
-                const double delta_code_m =
-                    current.corrected_pseudorange_m - previous.corrected_pseudorange_m;
-                const double code_phase_jump_m = std::abs(delta_carrier_m - delta_code_m);
-                if (config_.reject_tdcp_code_phase_jump &&
-                    config_.tdcp_code_phase_jump_threshold_m > 0.0 &&
-                    code_phase_jump_m > config_.tdcp_code_phase_jump_threshold_m) {
-                    ++problem.diagnostics.tdcp_rejected_code_phase_jump;
-                    continue;
-                }
-
-                TimeDifferencedCarrierFactor factor;
-                factor.previous_epoch_index = epoch_index - 1;
-                factor.current_epoch_index = epoch_index;
-                factor.satellite = current.satellite;
-                factor.signal = current.signal;
-                factor.previous_satellite_position_ecef = previous.satellite_position_ecef;
-                factor.current_satellite_position_ecef = current.satellite_position_ecef;
-                factor.delta_carrier_m = delta_carrier_m;
-                factor.sigma_m = sigma;
-                factor.dt_s = dt;
-                problem.tdcp_factors.push_back(factor);
-            }
-        }
-    }
-
-    if (config_.use_double_difference_factors) {
-        for (std::size_t epoch_index = 0;
-             epoch_index < dd_reference_carrier_by_problem_epoch.size();
-             ++epoch_index) {
-            for (const auto& [key, carrier] :
-                 dd_reference_carrier_by_problem_epoch[epoch_index]) {
-                CarrierPhaseFactor reference_observation;
-                reference_observation.epoch_index = epoch_index;
-                reference_observation.satellite = carrier.satellite;
-                reference_observation.clock_group =
-                    clockBiasGroup(carrier.satellite.system);
-                reference_observation.signal = carrier.signal;
-                reference_observation.satellite_position_ecef =
-                    carrier.satellite_position_ecef;
-                reference_observation.corrected_pseudorange_m =
-                    carrier.corrected_pseudorange_m;
-                reference_observation.corrected_carrier_m =
-                    carrier.corrected_carrier_m;
-                reference_observation.wavelength_m = carrier.wavelength_m;
-                reference_observation.sigma_m = carrier.sigma_m;
-                reference_observation.elevation_rad = carrier.elevation_rad;
-                reference_observation.has_carrier_phase = carrier.has_carrier_phase;
-                reference_observation.loss_of_lock = carrier.loss_of_lock;
-                reference_observation.has_doppler_residual =
-                    carrier.has_doppler_residual;
-                reference_observation.doppler_residual_mps =
-                    carrier.doppler_residual_mps;
-                reference_observation.doppler_sigma_mps =
-                    carrier.doppler_sigma_mps;
-                reference_observation.los = carrier.los;
-                reference_observation.model_debug = carrier.model_debug;
-                problem.double_difference_reference_observations.push_back(
-                    reference_observation);
-            }
-        }
-    }
-
-    return problem;
-}
-
-FGOProcessor::FGOProblem FGOProcessor::buildDoubleDifferenceProblem(
-    const std::vector<ObservationData>& rover_epochs,
-    const std::vector<ObservationData>& base_epochs,
-    const NavigationData& nav,
-    const Vector3d& base_position_ecef) const {
-    FGOProblem problem = buildPseudorangeProblem(rover_epochs, nav);
-    if (!config_.use_double_difference_factors ||
-        problem.carrier_observations.empty() ||
-        base_epochs.empty() ||
-        base_position_ecef.norm() <= 1e6) {
-        return problem;
-    }
-
-    std::map<std::size_t, std::vector<const CarrierPhaseFactor*>>
-        rover_carriers_by_epoch;
-    for (const auto& factor : problem.carrier_observations) {
-        if (factor.epoch_index < problem.epochs.size()) {
-            rover_carriers_by_epoch[factor.epoch_index].push_back(&factor);
-        }
-    }
-    std::map<std::size_t, std::vector<const CarrierPhaseFactor*>>
-        rover_pseudoranges_by_epoch;
-    for (const auto& factor :
-         problem.double_difference_pseudorange_observations) {
-        if (factor.epoch_index < problem.epochs.size()) {
-            rover_pseudoranges_by_epoch[factor.epoch_index].push_back(&factor);
-        }
-    }
-    std::map<std::size_t, std::vector<const CarrierPhaseFactor*>>
-        rover_reference_carriers_by_epoch;
-    for (const auto& factor : problem.double_difference_reference_observations) {
-        if (factor.epoch_index < problem.epochs.size()) {
-            rover_reference_carriers_by_epoch[factor.epoch_index].push_back(&factor);
-        }
-    }
-
-    // MF hygiene: the base receiver gets its own per-(system, secondary band)
-    // code table, mirroring the rover-side table in buildPseudorangeProblem.
-    SecondaryCodeTable base_secondary_codes;
-    const SecondaryCodeTable* base_secondary_codes_ptr = nullptr;
-    if (config_.use_multi_frequency_double_difference &&
-        config_.use_double_difference_secondary_code_alignment &&
-        config_.use_multi_constellation) {
-        base_secondary_codes = buildSecondaryCodeTable(base_epochs);
-        base_secondary_codes_ptr = &base_secondary_codes;
-    }
-
-    std::size_t base_cursor = 0;
-    const double match_tolerance = std::max(0.0, config_.base_epoch_match_tolerance_s);
-    const double pseudorange_sigma =
-        std::max(1e-3, config_.double_difference_pseudorange_sigma_m);
-    const double carrier_sigma =
-        std::max(1e-4, config_.double_difference_carrier_sigma_m);
-    // Elevation-dependent DD sigma ("varerr", FGOConfig::use_elevation_
-    // dependent_sigma): el_min_rad is the pair-elevation floor (reference:
-    // np.radians(max(1.0, cfg.el_mask_deg))); epoch_dt_s is the rover
-    // epoch-to-epoch interval fed to the clock-stability term, persisted
-    // across the epoch loop below and only updated on a positive delta --
-    // mirroring the reference's tc._epoch_dt / _update_epoch_dt exactly
-    // (state carried across epochs, default 0.2 s until the first update;
-    // see the knob's doc comment in fgo.hpp for why 0.2 s is also the right
-    // fallback for this dataset).
-    const double el_min_rad = M_PI / 180.0 * std::max(1.0, config_.min_elevation_deg);
-    double epoch_dt_s = 0.2;
-    const double max_segment_gap = std::max(0.0, config_.max_tdcp_gap_s);
-    std::map<DoubleDifferenceAmbiguityKey, ActiveCarrierSegment>
-        active_dd_segments;
-    std::map<DoubleDifferenceSegmentKey, std::size_t> next_dd_segment_indices;
-    using DdFactorKey =
-        std::tuple<std::size_t, SatelliteId, SatelliteId, SignalType>;
-    std::map<SingleDifferenceReferenceKey, SatelliteId> sd_reference_by_group;
-    std::map<SingleDifferenceDefaultReferenceKey, std::map<SatelliteId, std::size_t>>
-        sd_default_reference_counts;
-
-    // --- CMC screening state (function-scope, persists across the epoch
-    // loop below; see FGOConfig::use_code_minus_carrier_screening). ---
-    std::map<CarrierKey, CmcState> cmc_states;
-    // Last rover-side (single-receiver) ambiguity_index seen per (satellite,
-    // signal); a change here (cycle slip / loss-of-lock / outage causing the
-    // rover carrier arc in buildPseudorangeProblem to restart) is this port's
-    // signal to also reset the CMC baseline for that key (deviation from the
-    // reference documented on the config knob).
-    std::map<CarrierKey, std::size_t> cmc_last_rover_ambiguity_index;
-    std::size_t cmc_jump_reset_total = 0;
-    using GeometryFreePairKey =
-        std::tuple<SatelliteId, SignalType, SignalType>;
-    std::map<GeometryFreePairKey, GeometryFreeSlipState>
-        geometry_free_slip_states;
-    std::map<CarrierKey, GeometryFreePendingReset>
-        geometry_free_pending_resets;
-    std::size_t geometry_free_slip_reset_total = 0;
-    std::size_t cmc_level_exclusion_total = 0;
-    // FGOConfig::cmc_aware_reference_selection: count of DD reference
-    // picks steered away from a CMC-level-excluded candidate this run.
-    std::size_t cmc_ref_avoided_total = 0;
-
-    for (std::size_t epoch_index = 0; epoch_index < problem.epochs.size(); ++epoch_index) {
-        // Reference: tc._update_epoch_dt(obs) runs unconditionally at the top
-        // of every epoch's processing, before any DD-eligibility gates below
-        // -- mirrored here the same way (only meaningful when varerr is on).
-        if (config_.use_elevation_dependent_sigma && epoch_index > 0) {
-            const double dt = problem.epochs[epoch_index].time -
-                              problem.epochs[epoch_index - 1].time;
-            if (dt > 0.0) {
-                epoch_dt_s = dt;
-            }
-        }
-        const auto rover_it = rover_carriers_by_epoch.find(epoch_index);
-        const auto rover_pseudorange_it =
-            rover_pseudoranges_by_epoch.find(epoch_index);
-        const auto rover_reference_it =
-            rover_reference_carriers_by_epoch.find(epoch_index);
-        if ((rover_it == rover_carriers_by_epoch.end() ||
-             rover_it->second.empty()) &&
-            (rover_pseudorange_it == rover_pseudoranges_by_epoch.end() ||
-             rover_pseudorange_it->second.empty())) {
-            continue;
-        }
-
-        ObservationData matched_base_epoch;
-        bool interpolated_base_epoch = false;
-        const bool has_base_epoch =
-            findMatchedBaseEpoch(base_epochs,
-                                 problem.epochs[epoch_index].time,
-                                 base_cursor,
-                                 match_tolerance,
-                                 config_.base_interpolation_max_gap_s,
-                                 matched_base_epoch,
-                                 interpolated_base_epoch);
-        if (!has_base_epoch) {
-            problem.diagnostics.double_difference_rejected_no_base_epoch +=
-                rover_it == rover_carriers_by_epoch.end()
-                    ? 0
-                    : rover_it->second.size();
-            continue;
-        }
-        ++problem.diagnostics.double_difference_matched_base_epochs;
-        if (interpolated_base_epoch) {
-            ++problem.diagnostics.double_difference_interpolated_base_epochs;
-        }
-
-        std::map<std::pair<GNSSSystem, SignalType>, std::vector<const CarrierPhaseFactor*>>
-            grouped_rover_pseudoranges;
-        std::map<std::pair<GNSSSystem, SignalType>, std::vector<const CarrierPhaseFactor*>>
-            grouped_rover_carriers;
-        std::map<CarrierKey, const CarrierPhaseFactor*> rover_references_by_key;
-        if (rover_reference_it != rover_reference_carriers_by_epoch.end()) {
-            for (const auto* reference_factor : rover_reference_it->second) {
-                rover_references_by_key[{reference_factor->satellite,
-                                          reference_factor->signal}] =
-                    reference_factor;
-            }
-        }
-        std::map<std::pair<GNSSSystem, SignalType>,
-                 std::vector<const PreparedCarrierObservation*>>
-            grouped_reference_observations;
-        const double base_min_snr_dbhz =
-            config_.double_difference_base_min_snr_dbhz >= 0.0
-                ? config_.double_difference_base_min_snr_dbhz
-                : config_.min_snr_dbhz;
-        const double reference_min_snr_dbhz =
-            config_.double_difference_reference_min_snr_dbhz >= 0.0
-                ? config_.double_difference_reference_min_snr_dbhz
-                : config_.min_snr_dbhz;
-        const auto base_carriers =
-            prepareCarrierObservationsForReceiver(
-                matched_base_epoch,
-                nav,
-                base_position_ecef,
-                config_,
-                base_min_snr_dbhz,
-                true,
-                false,
-                base_secondary_codes_ptr);
-        const auto base_pseudorange_observations =
-            prepareCarrierObservationsForReceiver(
-                matched_base_epoch,
-                nav,
-                base_position_ecef,
-                config_,
-                base_min_snr_dbhz,
-                false,
-                false,
-                base_secondary_codes_ptr);
-        const auto base_reference_observations =
-            prepareCarrierObservationsForReceiver(
-                matched_base_epoch,
-                nav,
-                base_position_ecef,
-                config_,
-                reference_min_snr_dbhz,
-                false,
-                false,
-                base_secondary_codes_ptr);
-
-        // --- CMC screening pre-pass (this epoch) ---
-        //
-        // Runs once per (satellite, signal) that has usable carrier at BOTH
-        // rover and base this epoch (same requirement as the reference:
-        // pr/cp nonzero on both sides), independent of which satellite ends
-        // up the DD reference for its (system, signal) group below. Results
-        // gate the PR-factor loop (level exclusion, which -- because the CP
-        // loop only builds a factor when a matching PR factor already exists
-        // -- transitively also excludes the CP factor for the same epoch,
-        // matching the reference's single `continue` before either) and the
-        // CP loop's ambiguity-arc bookkeeping (jump forces a new arc).
-        std::set<CarrierKey> cmc_level_exclude_this_epoch;
-        std::set<CarrierKey> cmc_jump_reset_this_epoch;
-        std::set<CarrierKey> geometry_free_reset_this_epoch;
-        if (config_.use_code_minus_carrier_screening &&
-            rover_it != rover_carriers_by_epoch.end()) {
-            for (const auto* rover_factor : rover_it->second) {
-                const CarrierKey key{rover_factor->satellite, rover_factor->signal};
-                const auto base_it = base_carriers.find(key);
-                if (base_it == base_carriers.end()) {
-                    continue;
-                }
-                const auto& base_carrier = base_it->second;
-                const double pr_rover = rover_factor->model_debug.raw_pseudorange_m;
-                const double cp_rover = rover_factor->model_debug.raw_carrier_m;
-                const double pr_base = base_carrier.model_debug.raw_pseudorange_m;
-                const double cp_base = base_carrier.model_debug.raw_carrier_m;
-                if (pr_rover == 0.0 || cp_rover == 0.0 || pr_base == 0.0 ||
-                    cp_base == 0.0) {
-                    continue;
-                }
-                const double cmc = (pr_rover - pr_base) - (cp_rover - cp_base);
-
-                // Deviation from the reference: any rover-side arc restart
-                // (cycle slip / loss-of-lock / outage -- surfaced here as a
-                // change in the rover carrier's own ambiguity_index) resets
-                // this key's CMC baseline/prev/warmup state, since CMC
-                // embeds a -wavelength*N term that a new ambiguity
-                // invalidates.
-                const auto last_arc_it =
-                    cmc_last_rover_ambiguity_index.find(key);
-                const bool rover_arc_restarted =
-                    last_arc_it == cmc_last_rover_ambiguity_index.end() ||
-                    last_arc_it->second != rover_factor->ambiguity_index;
-                cmc_last_rover_ambiguity_index[key] = rover_factor->ambiguity_index;
-
-                CmcState& state = cmc_states[key];
-                if (rover_arc_restarted) {
-                    state = CmcState{};
-                } else if (state.has_prev &&
-                           std::abs(cmc - state.prev_cmc_m) >
-                               config_.code_minus_carrier_jump_threshold_m) {
-                    cmc_jump_reset_this_epoch.insert(key);
-                    ++cmc_jump_reset_total;
-                    // Same reasoning as above: the jump itself is a fresh
-                    // arc from this epoch on, so the baseline resets too.
-                    state = CmcState{};
-                }
-
-                state.prev_cmc_m = cmc;
-                state.has_prev = true;
-
-                const double level_threshold =
-                    config_.code_minus_carrier_level_threshold_m;
-                if (level_threshold > 0.0) {
-                    if (!state.has_baseline) {
-                        state.baseline_m = cmc;
-                        state.warmup_count = 1;
-                        state.has_baseline = true;
-                    } else if (state.warmup_count <
-                               config_.code_minus_carrier_warmup_epochs) {
-                        state.baseline_m =
-                            (state.baseline_m * state.warmup_count + cmc) /
-                            (state.warmup_count + 1);
-                        ++state.warmup_count;
-                    } else if (std::abs(cmc - state.baseline_m) > level_threshold) {
-                        cmc_level_exclude_this_epoch.insert(key);
-                        ++cmc_level_exclusion_total;
-                    } else {
-                        const double alpha = config_.code_minus_carrier_baseline_alpha;
-                        state.baseline_m =
-                            (1.0 - alpha) * state.baseline_m + alpha * cmc;
-                    }
-                }
-            }
-        }
-
-        if (config_.use_geometry_free_cycle_slip_reset &&
-            rover_it != rover_carriers_by_epoch.end()) {
-            using GeometryFreeSample =
-                std::pair<const CarrierPhaseFactor*,
-                          const PreparedCarrierObservation*>;
-            std::map<SatelliteId, std::map<SignalType, GeometryFreeSample>>
-                samples_by_satellite;
-            std::map<CarrierKey, GeometryFreeSample> samples_by_key;
-            for (const auto* rover_factor : rover_it->second) {
-                const CarrierKey key{rover_factor->satellite,
-                                     rover_factor->signal};
-                const auto base_it = base_carriers.find(key);
-                if (base_it != base_carriers.end()) {
-                    const GeometryFreeSample sample{rover_factor,
-                                                    &base_it->second};
-                    samples_by_satellite[rover_factor->satellite]
-                                        [rover_factor->signal] = sample;
-                    samples_by_key[key] = sample;
-                }
-            }
-            const double confirmation_s = std::max(
-                0.0, config_.geometry_free_cycle_slip_confirmation_s);
-            for (auto pending = geometry_free_pending_resets.begin();
-                 pending != geometry_free_pending_resets.end();) {
-                const auto sample = samples_by_key.find(pending->first);
-                if (sample == samples_by_key.end() ||
-                    sample->second.first->ambiguity_index !=
-                        pending->second.rover_ambiguity_index) {
-                    pending = geometry_free_pending_resets.erase(pending);
-                    continue;
-                }
-                const double age_s = problem.epochs[epoch_index].time -
-                                     pending->second.time;
-                if (age_s >= confirmation_s) {
-                    geometry_free_reset_this_epoch.insert(pending->first);
-                    ++geometry_free_slip_reset_total;
-                    pending = geometry_free_pending_resets.erase(pending);
-                    continue;
-                }
-                ++pending;
-            }
-            std::set<CarrierKey> slip_bands_this_epoch;
-            for (const auto& [satellite, samples_by_signal] :
-                 samples_by_satellite) {
-                std::vector<GeometryFreeSample> samples;
-                samples.reserve(samples_by_signal.size());
-                for (const auto& [signal, sample] : samples_by_signal) {
-                    (void)signal;
-                    samples.push_back(sample);
-                }
-                for (std::size_t first = 0; first < samples.size(); ++first) {
-                    for (std::size_t second = first + 1;
-                         second < samples.size(); ++second) {
-                        const auto* first_rover = samples[first].first;
-                        const auto* second_rover = samples[second].first;
-                        const auto* first_base = samples[first].second;
-                        const auto* second_base = samples[second].second;
-                        const double first_sd_phase_m =
-                            first_rover->model_debug.raw_carrier_m -
-                            first_base->model_debug.raw_carrier_m;
-                        const double second_sd_phase_m =
-                            second_rover->model_debug.raw_carrier_m -
-                            second_base->model_debug.raw_carrier_m;
-                        const double geometry_free_m =
-                            first_sd_phase_m - second_sd_phase_m;
-                        const GeometryFreePairKey pair_key{
-                            satellite, first_rover->signal,
-                            second_rover->signal};
-                        const auto previous =
-                            geometry_free_slip_states.find(pair_key);
-                        if (previous != geometry_free_slip_states.end()) {
-                            const double dt = problem.epochs[epoch_index].time -
-                                              previous->second.time;
-                            const bool same_rover_arcs =
-                                previous->second.first_rover_ambiguity_index ==
-                                    first_rover->ambiguity_index &&
-                                previous->second.second_rover_ambiguity_index ==
-                                    second_rover->ambiguity_index;
-                            if (same_rover_arcs && dt > 0.0 &&
-                                (max_segment_gap <= 0.0 ||
-                                 dt <= max_segment_gap) &&
-                                std::abs(geometry_free_m -
-                                         previous->second.geometry_free_m) >
-                                    std::max(
-                                        0.0,
-                                        config_.geometry_free_cycle_slip_threshold_m)) {
-                                const CarrierKey first_key{
-                                    satellite, first_rover->signal};
-                                const CarrierKey second_key{
-                                    satellite, second_rover->signal};
-                                slip_bands_this_epoch.insert(first_key);
-                                slip_bands_this_epoch.insert(second_key);
-                            }
-                        }
-                        geometry_free_slip_states[pair_key] = {
-                            problem.epochs[epoch_index].time,
-                            geometry_free_m,
-                            first_rover->ambiguity_index,
-                            second_rover->ambiguity_index};
-                    }
-                }
-            }
-            for (const auto& key : slip_bands_this_epoch) {
-                if (geometry_free_reset_this_epoch.count(key) != 0) {
-                    continue;
-                }
-                const auto sample = samples_by_key.find(key);
-                if (sample == samples_by_key.end()) {
-                    continue;
-                }
-                const auto pending = geometry_free_pending_resets.find(key);
-                if (pending != geometry_free_pending_resets.end()) {
-                    // A second discontinuity before confirmation is
-                    // ambiguous (often an isolated outlier returning to its
-                    // former level), so fail closed instead of restarting the
-                    // timer.
-                    geometry_free_pending_resets.erase(pending);
-                    continue;
-                }
-                geometry_free_pending_resets[key] = {
-                    problem.epochs[epoch_index].time,
-                    sample->second.first->ambiguity_index};
-            }
-        }
-
-        if (rover_pseudorange_it != rover_pseudoranges_by_epoch.end()) {
-            for (const auto* rover_factor : rover_pseudorange_it->second) {
-                const CarrierKey key{rover_factor->satellite,
-                                     rover_factor->signal};
-                if (base_pseudorange_observations.find(key) ==
-                    base_pseudorange_observations.end()) {
-                    continue;
-                }
-                grouped_rover_pseudoranges[{rover_factor->satellite.system,
-                                            rover_factor->signal}]
-                    .push_back(rover_factor);
-            }
-        }
-        if (rover_it != rover_carriers_by_epoch.end()) {
-            for (const auto* rover_factor : rover_it->second) {
-                const CarrierKey key{rover_factor->satellite,
-                                     rover_factor->signal};
-                if (base_carriers.find(key) == base_carriers.end()) {
-                    continue;
-                }
-                grouped_rover_carriers[{rover_factor->satellite.system,
-                                        rover_factor->signal}]
-                    .push_back(rover_factor);
-            }
-        }
-        for (const auto& [key, reference_observation] :
-             base_reference_observations) {
-            grouped_reference_observations[{reference_observation.satellite.system,
-                                            reference_observation.signal}]
-                .push_back(&reference_observation);
-        }
-
-        const auto select_reference =
-            [&](const std::pair<GNSSSystem, SignalType>& group_key,
-                std::size_t rejected_count,
-                const CarrierPhaseFactor*& reference,
-                const PreparedCarrierObservation*& base_reference) {
-                if (group_key.first == GNSSSystem::GLONASS) {
-                    problem.diagnostics.double_difference_rejected_no_reference +=
-                        rejected_count;
-                    return false;
-                }
-
-                const auto reference_group_it =
-                    grouped_reference_observations.find(group_key);
-                if (reference_group_it == grouped_reference_observations.end() ||
-                    reference_group_it->second.empty()) {
-                    problem.diagnostics.double_difference_rejected_no_reference +=
-                        rejected_count;
-                    return false;
-                }
-
-                const PreparedCarrierObservation* base_reference_selection =
-                    *std::max_element(
-                        reference_group_it->second.begin(),
-                        reference_group_it->second.end(),
-                        [](const PreparedCarrierObservation* lhs,
-                           const PreparedCarrierObservation* rhs) {
-                            return lhs->elevation_rad < rhs->elevation_rad;
-                        });
-
-                // --- CMC-aware reference selection
-                // (FGOConfig::cmc_aware_reference_selection) ---
-                // The plain max-elevation pick above can land on a satellite
-                // the CMC screening pre-pass above already flagged as
-                // sustained multipath/NLOS this epoch
-                // (cmc_level_exclude_this_epoch). Every DD pair in the group
-                // is formed against this ONE reference, so a multipath-
-                // biased reference poisons every DD residual in the group at
-                // once (see the config knob's doc comment for the tokyo
-                // run1 episode that motivated this). When the knob is on and
-                // CMC screening is active, prefer the highest-elevation
-                // NON-excluded candidate instead; if every candidate in the
-                // group is CMC-excluded this epoch, keep the original
-                // max-elevation choice so the group is never dropped.
-                if (config_.cmc_aware_reference_selection &&
-                    config_.use_code_minus_carrier_screening &&
-                    !cmc_level_exclude_this_epoch.empty()) {
-                    const CarrierKey default_key{
-                        base_reference_selection->satellite,
-                        base_reference_selection->signal};
-                    if (cmc_level_exclude_this_epoch.count(default_key) > 0) {
-                        const PreparedCarrierObservation* best_clean = nullptr;
-                        for (const auto* candidate : reference_group_it->second) {
-                            const CarrierKey candidate_key{candidate->satellite,
-                                                            candidate->signal};
-                            if (cmc_level_exclude_this_epoch.count(candidate_key) >
-                                0) {
-                                continue;
-                            }
-                            if (!best_clean || candidate->elevation_rad >
-                                                    best_clean->elevation_rad) {
-                                best_clean = candidate;
-                            }
-                        }
-                        if (best_clean) {
-                            base_reference_selection = best_clean;
-                            ++cmc_ref_avoided_total;
-                        }
-                        // else: every candidate in the group is CMC-excluded
-                        // -- fall back to the original max-elevation choice
-                        // (base_reference_selection unchanged) rather than
-                        // dropping the whole group.
-                    }
-                }
-
-                const CarrierKey reference_key{
-                    base_reference_selection->satellite,
-                    base_reference_selection->signal};
-                const auto rover_reference_by_key_it =
-                    rover_references_by_key.find(reference_key);
-                if (rover_reference_by_key_it == rover_references_by_key.end()) {
-                    problem.diagnostics.double_difference_rejected_no_reference +=
-                        rejected_count;
-                    return false;
-                }
-                reference = rover_reference_by_key_it->second;
-                base_reference = base_reference_selection;
-                return true;
-            };
-
-        std::map<DdFactorKey, DoubleDifferencePseudorangeFactor>
-            pseudorange_factors_by_key;
-        // Includes CMC-level-excluded code observations solely for carrier
-        // ambiguity initialization when code-only exclusion is requested.
-        std::map<DdFactorKey, DoubleDifferencePseudorangeFactor>
-            pseudorange_seed_factors_by_key;
-        for (const auto& [group_key, group] : grouped_rover_pseudoranges) {
-            if (group_key.first == GNSSSystem::GLONASS) {
-                problem.diagnostics.double_difference_rejected_no_reference +=
-                    group.size();
-                continue;
-            }
-
-            const CarrierPhaseFactor* reference = nullptr;
-            const PreparedCarrierObservation* base_reference_ptr = nullptr;
-            if (!select_reference(group_key,
-                                  group.size(),
-                                  reference,
-                                  base_reference_ptr)) {
-                continue;
-            }
-            const auto& base_reference = *base_reference_ptr;
-            for (const auto* satellite : group) {
-                if (satellite->satellite == reference->satellite &&
-                    satellite->signal == reference->signal) {
-                    continue;
-                }
-                const CarrierKey satellite_key{satellite->satellite, satellite->signal};
-                const bool cmc_level_excluded =
-                    cmc_level_exclude_this_epoch.count(satellite_key) > 0;
-                if (cmc_level_excluded &&
-                    !config_.code_minus_carrier_level_pseudorange_only) {
-                    // Sustained CMC multipath this epoch: skip the DD
-                    // pseudorange factor. The CP loop below only builds a
-                    // carrier factor when it finds a matching entry in
-                    // pseudorange_factors_by_key, so omitting the PR factor
-                    // here transitively excludes the CP factor too (matches
-                    // the reference's single `continue` before either).
-                    continue;
-                }
-                const auto base_satellite_it =
-                    base_pseudorange_observations.find(satellite_key);
-                if (base_satellite_it == base_pseudorange_observations.end()) {
-                    continue;
-                }
-
-                const auto& base_satellite = base_satellite_it->second;
-                const double satellite_sin_el =
-                    std::max(0.1, std::sin(satellite->elevation_rad));
-                const double satellite_sqrt_sin_el =
-                    std::sqrt(satellite_sin_el);
-
-                DoubleDifferencePseudorangeFactor pseudorange_factor;
-                pseudorange_factor.epoch_index = epoch_index;
-                pseudorange_factor.satellite = satellite->satellite;
-                pseudorange_factor.reference_satellite = reference->satellite;
-                pseudorange_factor.signal = satellite->signal;
-                pseudorange_factor.rover_satellite_position_ecef =
-                    satellite->satellite_position_ecef;
-                pseudorange_factor.rover_reference_position_ecef =
-                    reference->satellite_position_ecef;
-                pseudorange_factor.base_satellite_position_ecef =
-                    base_satellite.satellite_position_ecef;
-                pseudorange_factor.base_reference_position_ecef =
-                    base_reference.satellite_position_ecef;
-                pseudorange_factor.base_position_ecef = base_position_ecef;
-                pseudorange_factor.observed_dd_pseudorange_m =
-                    (satellite->corrected_pseudorange_m -
-                     base_satellite.corrected_pseudorange_m) -
-                    (reference->corrected_pseudorange_m -
-                     base_reference.corrected_pseudorange_m);
-                const double pseudorange_band_scale =
-                    signal_policy::isSecondarySignal(satellite->satellite.system,
-                                                     satellite->signal)
-                        ? std::max(1.0,
-                                   config_
-                                       .double_difference_secondary_pseudorange_sigma_scale)
-                        : 1.0;
-                // Base DD-PR sigma: varerr (FGOConfig::use_elevation_
-                // dependent_sigma) when enabled, else the pre-existing flat/
-                // elevation-power fallback -- see the knob's doc comment in
-                // fgo.hpp for the full rationale/composition order. The
-                // secondary-band scale above still multiplies on top either
-                // way (unchanged).
-                double pseudorange_base_sigma = pseudorange_sigma / satellite_sqrt_sin_el;
-                if (config_.use_elevation_dependent_sigma) {
-                    const double el_pair_rad = std::max(
-                        std::min(reference->elevation_rad, satellite->elevation_rad),
-                        el_min_rad);
-                    pseudorange_base_sigma = varerrDdSigma(
-                        /*is_pseudorange=*/true, el_pair_rad, epoch_dt_s, config_);
-                }
-                pseudorange_factor.sigma_m =
-                    pseudorange_band_scale * pseudorange_base_sigma;
-                pseudorange_factor.elevation_rad = satellite->elevation_rad;
-                pseudorange_factor.rover_satellite_model =
-                    satellite->model_debug;
-                pseudorange_factor.rover_reference_model =
-                    reference->model_debug;
-                pseudorange_factor.base_satellite_model =
-                    base_satellite.model_debug;
-                pseudorange_factor.base_reference_model =
-                    base_reference.model_debug;
-                const DdFactorKey factor_key{
-                    epoch_index,
-                    satellite->satellite,
-                    reference->satellite,
-                    satellite->signal,
-                };
-                pseudorange_seed_factors_by_key[factor_key] = pseudorange_factor;
-                if (!cmc_level_excluded) {
-                    problem.double_difference_pseudorange_factors.push_back(
-                        pseudorange_factor);
-                    pseudorange_factors_by_key[factor_key] = pseudorange_factor;
-                }
-            }
-        }
-
-        for (const auto& [group_key, group] : grouped_rover_carriers) {
-            const CarrierPhaseFactor* reference = nullptr;
-            const PreparedCarrierObservation* base_reference_ptr = nullptr;
-            if (!select_reference(group_key,
-                                  group.size(),
-                                  reference,
-                                  base_reference_ptr)) {
-                continue;
-            }
-            const auto& base_reference = *base_reference_ptr;
-
-            for (const auto* satellite : group) {
-                if (satellite->satellite == reference->satellite &&
-                    satellite->signal == reference->signal) {
-                    continue;
-                }
-                const CarrierKey satellite_key{satellite->satellite, satellite->signal};
-                const auto base_satellite_it = base_carriers.find(satellite_key);
-                if (base_satellite_it == base_carriers.end()) {
-                    continue;
-                }
-                const DdFactorKey factor_key{
-                        epoch_index,
-                        satellite->satellite,
-                        reference->satellite,
-                        satellite->signal,
-                    };
-                const auto pseudorange_factor_it = pseudorange_factors_by_key.find(factor_key);
-                const DoubleDifferencePseudorangeFactor* pseudorange_factor_ptr =
-                    pseudorange_factor_it != pseudorange_factors_by_key.end()
-                        ? &pseudorange_factor_it->second
-                        : nullptr;
-                if (!pseudorange_factor_ptr &&
-                    config_.code_minus_carrier_level_pseudorange_only) {
-                    const auto seed_it = pseudorange_seed_factors_by_key.find(factor_key);
-                    if (seed_it != pseudorange_seed_factors_by_key.end()) {
-                        pseudorange_factor_ptr = &seed_it->second;
-                    }
-                }
-                if (!pseudorange_factor_ptr) {
-                    // Sustained CMC multipath dropped this satellite's DD-PR
-                    // factor at problem-build time (and, transitively, this
-                    // DD-CP row -- code_minus_carrier_level_pseudorange_only
-                    // is false, else pseudorange_factor_ptr would have come
-                    // from the seed map above). Retain a minimal excluded
-                    // row -- geometry + observed DD carrier, no ambiguity/
-                    // segment tracking (ambiguity_index is the sentinel
-                    // max()) -- solely so the surplus-satellite independent
-                    // integrity validator (FGOConfig::use_surplus_
-                    // satellite_validation) can see observations that never
-                    // entered ambiguity resolution at all. NEVER added to
-                    // double_difference_carrier_factors / the solved graph.
-                    if (cmc_level_exclude_this_epoch.count(satellite_key) > 0 &&
-                        reference->has_carrier_phase && base_reference.has_carrier_phase) {
-                        const auto& base_satellite_excl = base_satellite_it->second;
-                        DoubleDifferenceCarrierFactor excluded;
-                        excluded.epoch_index = epoch_index;
-                        excluded.use_ambiguity_difference = false;
-                        excluded.ambiguity_index = std::numeric_limits<std::size_t>::max();
-                        excluded.satellite = satellite->satellite;
-                        excluded.reference_satellite = reference->satellite;
-                        excluded.signal = satellite->signal;
-                        excluded.rover_satellite_position_ecef = satellite->satellite_position_ecef;
-                        excluded.rover_reference_position_ecef = reference->satellite_position_ecef;
-                        excluded.base_satellite_position_ecef =
-                            base_satellite_excl.satellite_position_ecef;
-                        excluded.base_reference_position_ecef =
-                            base_reference.satellite_position_ecef;
-                        excluded.base_position_ecef = base_position_ecef;
-                        excluded.observed_dd_carrier_m =
-                            (satellite->corrected_carrier_m -
-                             base_satellite_excl.corrected_carrier_m) -
-                            (reference->corrected_carrier_m -
-                             base_reference.corrected_carrier_m);
-                        excluded.sigma_m = carrier_sigma;  // unused by the surplus validator
-                        excluded.elevation_rad = satellite->elevation_rad;
-                        problem.excluded_double_difference_carrier_factors.push_back(excluded);
-                    }
-                    continue;
-                }
-                ++problem.diagnostics.double_difference_candidate_pairs;
-
-                const auto& base_satellite = base_satellite_it->second;
-                const auto& pseudorange_factor = *pseudorange_factor_ptr;
-                const double satellite_sin_el =
-                    std::max(0.1, std::sin(satellite->elevation_rad));
-                const double satellite_sqrt_sin_el =
-                    std::sqrt(satellite_sin_el);
-
-                if (!reference->has_carrier_phase ||
-                    !base_reference.has_carrier_phase) {
-                    continue;
-                }
-
-                DoubleDifferenceCarrierFactor factor;
-                factor.epoch_index = epoch_index;
-                factor.use_ambiguity_difference = false;
-                factor.satellite = satellite->satellite;
-                factor.reference_satellite = reference->satellite;
-                factor.signal = satellite->signal;
-                factor.rover_satellite_position_ecef =
-                    satellite->satellite_position_ecef;
-                factor.rover_reference_position_ecef =
-                    reference->satellite_position_ecef;
-                factor.base_satellite_position_ecef =
-                    base_satellite.satellite_position_ecef;
-                factor.base_reference_position_ecef =
-                    base_reference.satellite_position_ecef;
-                factor.base_position_ecef = base_position_ecef;
-                factor.observed_dd_carrier_m =
-                    (satellite->corrected_carrier_m -
-                     base_satellite.corrected_carrier_m) -
-                    (reference->corrected_carrier_m -
-                     base_reference.corrected_carrier_m);
-                const double carrier_band_scale =
-                    signal_policy::isSecondarySignal(satellite->satellite.system,
-                                                     satellite->signal)
-                        ? std::max(1.0,
-                                   config_
-                                       .double_difference_secondary_carrier_sigma_scale)
-                        : 1.0;
-                // Base DD-CP sigma: varerr when enabled, else the
-                // pre-existing flat/elevation-power fallback -- same source
-                // switch as the DD-PR site above (see that comment and the
-                // knob's doc comment in fgo.hpp). Secondary-band scale still
-                // multiplies on top either way.
-                double carrier_base_sigma = carrier_sigma / satellite_sqrt_sin_el;
-                if (config_.use_elevation_dependent_sigma) {
-                    const double el_pair_rad = std::max(
-                        std::min(reference->elevation_rad, satellite->elevation_rad),
-                        el_min_rad);
-                    carrier_base_sigma = varerrDdSigma(
-                        /*is_pseudorange=*/false, el_pair_rad, epoch_dt_s, config_);
-                }
-                factor.sigma_m = carrier_band_scale * carrier_base_sigma;
-                factor.elevation_rad = satellite->elevation_rad;
-                factor.rover_satellite_model = satellite->model_debug;
-                factor.rover_reference_model = reference->model_debug;
-                factor.base_satellite_model = base_satellite.model_debug;
-                factor.base_reference_model = base_reference.model_debug;
-
-                const Vector3d seed_position =
-                    problem.epochs[epoch_index].position_ecef;
-                const double rover_satellite_range =
-                    (factor.rover_satellite_position_ecef - seed_position).norm();
-                const double rover_reference_range =
-                    (factor.rover_reference_position_ecef - seed_position).norm();
-                const double base_satellite_range =
-                    (factor.base_satellite_position_ecef - base_position_ecef).norm();
-                const double base_reference_range =
-                    (factor.base_reference_position_ecef - base_position_ecef).norm();
-
-                const DoubleDifferenceAmbiguityKey ambiguity_key{
-                    satellite->satellite,
-                    reference->satellite,
-                    satellite->signal,
-                    satellite->ambiguity_index,
-                    0,
-                };
-                auto active_it = active_dd_segments.find(ambiguity_key);
-                bool start_new_segment =
-                    config_.reset_double_difference_ambiguities_each_epoch ||
-                    active_it == active_dd_segments.end();
-                if (!start_new_segment) {
-                    const double dt = problem.epochs[epoch_index].time -
-                                      problem.epochs[active_it->second.last_epoch_index].time;
-                    if (dt <= 0.0 ||
-                        (max_segment_gap > 0.0 && dt > max_segment_gap)) {
-                        start_new_segment = true;
-                    }
-                }
-                if (cmc_jump_reset_this_epoch.count(
-                        CarrierKey{satellite->satellite, satellite->signal}) > 0) {
-                    // CMC jump: force the same arc break the loss-of-lock /
-                    // gap checks above already perform (multipath jump
-                    // breaks the integer ambiguity).
-                    start_new_segment = true;
-                }
-                if (geometry_free_reset_this_epoch.count(
-                        CarrierKey{satellite->satellite, satellite->signal}) > 0 ||
-                    geometry_free_reset_this_epoch.count(
-                        CarrierKey{reference->satellite, reference->signal}) > 0) {
-                    start_new_segment = true;
-                }
-
-                if (start_new_segment) {
-                    AmbiguityState ambiguity;
-                    ambiguity.satellite = satellite->satellite;
-                    ambiguity.reference_satellite = reference->satellite;
-                    ambiguity.signal = satellite->signal;
-                    ambiguity.is_double_difference = true;
-                    ambiguity.segment_index =
-                        next_dd_segment_indices[DoubleDifferenceSegmentKey{
-                            satellite->satellite,
-                            reference->satellite,
-                            satellite->signal,
-                        }]++;
-                    ambiguity.wavelength_m = satellite->wavelength_m;
-                    const double dd_geometry_at_seed =
-                        (rover_satellite_range - base_satellite_range) -
-                        (rover_reference_range - base_reference_range);
-                    const double carrier_residual_at_seed =
-                        factor.observed_dd_carrier_m - dd_geometry_at_seed;
-                    const double pseudorange_residual_at_seed =
-                        pseudorange_factor.observed_dd_pseudorange_m -
-                        dd_geometry_at_seed;
-                    ambiguity.initial_ambiguity_m =
-                        carrier_residual_at_seed -
-                        pseudorange_residual_at_seed;
-                    active_it =
-                        active_dd_segments
-                            .insert_or_assign(
-                                ambiguity_key,
-                                ActiveCarrierSegment{problem.ambiguity_states.size(),
-                                                     epoch_index})
-                            .first;
-                    problem.ambiguity_states.push_back(ambiguity);
-                } else {
-                    active_it->second.last_epoch_index = epoch_index;
-                }
-
-                factor.ambiguity_index = active_it->second.ambiguity_index;
-                problem.double_difference_carrier_factors.push_back(factor);
-            }
-        }
-    }
-
-    for (const auto& factor : problem.double_difference_pseudorange_factors) {
-        sd_reference_by_group[SingleDifferenceReferenceKey{
-            factor.epoch_index,
-            factor.satellite.system,
-            factor.signal,
-        }] = factor.reference_satellite;
-        ++sd_default_reference_counts[SingleDifferenceDefaultReferenceKey{
-            factor.satellite.system,
-            factor.signal,
-        }][factor.reference_satellite];
-    }
-    for (const auto& factor : problem.double_difference_carrier_factors) {
-        sd_reference_by_group[SingleDifferenceReferenceKey{
-            factor.epoch_index,
-            factor.satellite.system,
-            factor.signal,
-        }] = factor.reference_satellite;
-        ++sd_default_reference_counts[SingleDifferenceDefaultReferenceKey{
-            factor.satellite.system,
-            factor.signal,
-        }][factor.reference_satellite];
-    }
-
-    if (config_.use_ambiguity_between_factors) {
-        using AmbiguityTrackKey = std::pair<SatelliteId, SignalType>;
-        std::map<AmbiguityTrackKey, const DoubleDifferenceCarrierFactor*>
-            previous_by_track;
-        for (const auto& factor : problem.double_difference_carrier_factors) {
-            if (factor.ambiguity_index >= problem.ambiguity_states.size()) {
-                continue;
-            }
-            const AmbiguityTrackKey track_key{factor.satellite, factor.signal};
-            const auto previous_it = previous_by_track.find(track_key);
-            if (previous_it != previous_by_track.end()) {
-                const auto* previous = previous_it->second;
-                const double dt =
-                    problem.epochs[factor.epoch_index].time -
-                    problem.epochs[previous->epoch_index].time;
-                if (factor.epoch_index == previous->epoch_index + 1 &&
-                    dt > 0.0 &&
-                    (config_.max_tdcp_gap_s <= 0.0 ||
-                     dt <= config_.max_tdcp_gap_s) &&
-                    previous->ambiguity_index != factor.ambiguity_index) {
-                    const auto& ambiguity =
-                        problem.ambiguity_states[factor.ambiguity_index];
-                    if (ambiguity.wavelength_m > 0.0) {
-                        AmbiguityBetweenFactor between;
-                        between.previous_epoch_index = previous->epoch_index;
-                        between.current_epoch_index = factor.epoch_index;
-                        between.previous_ambiguity_index =
-                            previous->ambiguity_index;
-                        between.current_ambiguity_index = factor.ambiguity_index;
-                        between.satellite = factor.satellite;
-                        between.signal = factor.signal;
-                        between.sigma_m =
-                            std::max(1e-9,
-                                     config_.ambiguity_between_sigma_cycles *
-                                         ambiguity.wavelength_m);
-                        problem.ambiguity_between_factors.push_back(between);
-                    }
-                }
-            }
-            previous_by_track[track_key] = &factor;
-        }
-    }
-
-    std::map<SingleDifferenceDefaultReferenceKey, SatelliteId>
-        sd_default_reference_by_group;
-    for (const auto& [group_key, counts] : sd_default_reference_counts) {
-        const auto best_it =
-            std::max_element(counts.begin(),
-                             counts.end(),
-                             [](const auto& lhs, const auto& rhs) {
-                                 return lhs.second < rhs.second;
-                             });
-        if (best_it != counts.end()) {
-            sd_default_reference_by_group[group_key] = best_it->first;
-        }
-    }
-
-    if (config_.use_single_difference_doppler_factors ||
-        config_.monitor_external_doppler_dr ||
-        config_.use_external_doppler_dr_validation ||
-        config_.monitor_candidate_integrity_witness ||
-        config_.use_single_difference_tdcp_factors) {
-        std::map<CarrierKey, SingleDifferenceCarrierResidual>
-            previous_sd_carrier_residuals;
-        for (std::size_t epoch_index = 0; epoch_index < problem.epochs.size();
-             ++epoch_index) {
-            std::map<CarrierKey, const CarrierPhaseFactor*> rover_by_key;
-            std::vector<const CarrierPhaseFactor*> rover_factor_rows;
-            const auto rover_pseudorange_it =
-                rover_pseudoranges_by_epoch.find(epoch_index);
-            if (rover_pseudorange_it != rover_pseudoranges_by_epoch.end()) {
-                for (const auto* rover : rover_pseudorange_it->second) {
-                    rover_by_key[{rover->satellite, rover->signal}] = rover;
-                    rover_factor_rows.push_back(rover);
-                }
-            }
-
-            std::map<CarrierKey, SingleDifferenceCarrierResidual>
-                current_sd_carrier_residuals;
-            for (const auto* rover : rover_factor_rows) {
-                const CarrierKey carrier_key{rover->satellite, rover->signal};
-                const auto reference_it =
-                    sd_reference_by_group.find(SingleDifferenceReferenceKey{
-                        epoch_index,
-                        rover->satellite.system,
-                        rover->signal,
-                    });
-                SatelliteId reference_satellite;
-                if (reference_it != sd_reference_by_group.end()) {
-                    reference_satellite = reference_it->second;
-                } else {
-                    const auto default_reference_it =
-                        sd_default_reference_by_group.find(
-                            SingleDifferenceDefaultReferenceKey{
-                                rover->satellite.system,
-                                rover->signal,
-                            });
-                    if (default_reference_it ==
-                            sd_default_reference_by_group.end() ||
-                        !(rover->satellite == default_reference_it->second)) {
-                        continue;
-                    }
-                    reference_satellite = default_reference_it->second;
-                }
-
-                const auto reference_model_it =
-                    rover_by_key.find({reference_satellite, rover->signal});
-                if (reference_model_it == rover_by_key.end()) {
-                    continue;
-                }
-                const auto* reference = reference_model_it->second;
-                // A satellite differenced with itself is identically zero
-                // and adds no information (but can distort factor counts and
-                // residual diagnostics).
-                if (rover->satellite == reference_satellite) {
-                    continue;
-                }
-                const Vector3d sd_los = rover->los - reference->los;
-
-                if ((config_.use_single_difference_doppler_factors ||
-                     config_.monitor_external_doppler_dr ||
-                     config_.use_external_doppler_dr_validation ||
-                     config_.monitor_candidate_integrity_witness) &&
-                    rover->has_doppler_residual &&
-                    reference->has_doppler_residual) {
-                    SingleDifferenceDopplerFactor factor;
-                    factor.epoch_index = epoch_index;
-                    factor.satellite = rover->satellite;
-                    factor.reference_satellite = reference_satellite;
-                    factor.signal = rover->signal;
-                    factor.los = sd_los;
-                    factor.residual_mps =
-                        rover->doppler_residual_mps -
-                        reference->doppler_residual_mps;
-                    factor.sigma_mps = std::hypot(rover->doppler_sigma_mps,
-                                                  reference->doppler_sigma_mps);
-                    factor.elevation_rad = rover->elevation_rad;
-                    if (std::isfinite(factor.residual_mps) &&
-                        std::isfinite(factor.sigma_mps) &&
-                        factor.sigma_mps > 0.0) {
-                        problem.single_difference_doppler_factors.push_back(factor);
-                    }
-                }
-
-                if (rover->has_carrier_phase &&
-                    reference->has_carrier_phase &&
-                    !rover->loss_of_lock &&
-                    !reference->loss_of_lock) {
-                    const double rover_carrier_residual =
-                        rover->corrected_carrier_m -
-                        rover->model_debug.geometric_range_m;
-                    const double reference_carrier_residual =
-                        reference->corrected_carrier_m -
-                        reference->model_debug.geometric_range_m;
-                    const double sd_carrier_residual =
-                        rover_carrier_residual - reference_carrier_residual;
-                    current_sd_carrier_residuals[carrier_key] =
-                        SingleDifferenceCarrierResidual{
-                            epoch_index,
-                            sd_carrier_residual,
-                            sd_los,
-                        };
-
-                    const auto previous_it =
-                        previous_sd_carrier_residuals.find(carrier_key);
-                    if (config_.use_single_difference_tdcp_factors &&
-                        previous_it != previous_sd_carrier_residuals.end()) {
-                        const double tdcp =
-                            sd_carrier_residual - previous_it->second.residual_m;
-                        if (std::isfinite(tdcp) && tdcp != 0.0) {
-                            SingleDifferenceTdcpFactor factor;
-                            factor.previous_epoch_index =
-                                previous_it->second.epoch_index;
-                            factor.current_epoch_index = epoch_index;
-                            factor.satellite = rover->satellite;
-                            factor.reference_satellite = reference_satellite;
-                            factor.signal = rover->signal;
-                            factor.previous_los = sd_los;
-                            factor.los = sd_los;
-                            factor.delta_carrier_m = tdcp;
-                            const double sin_el =
-                                std::max(0.1, std::sin(rover->elevation_rad));
-                            factor.sigma_m =
-                                std::max(1e-4,
-                                         config_.single_difference_tdcp_sigma_m /
-                                             std::sqrt(sin_el));
-                            factor.elevation_rad = rover->elevation_rad;
-                            problem.single_difference_tdcp_factors.push_back(factor);
-                        }
-                    }
-                }
-            }
-            previous_sd_carrier_residuals =
-                std::move(current_sd_carrier_residuals);
-        }
-    }
-
-    problem.diagnostics.code_minus_carrier_jump_resets = cmc_jump_reset_total;
-    problem.diagnostics.geometry_free_cycle_slip_resets =
-        geometry_free_slip_reset_total;
-    problem.diagnostics.code_minus_carrier_level_exclusions =
-        cmc_level_exclusion_total;
-    problem.diagnostics.cmc_ref_avoided_count = cmc_ref_avoided_total;
-
-    return problem;
-}
+using namespace fgo_internal;
 
 FGOProcessor::FGOResult FGOProcessor::optimize(
     const std::vector<ObservationData>& epochs,
@@ -2640,892 +38,6 @@ FGOProcessor::FGOResult FGOProcessor::optimize(
     return optimizeProblem(buildPseudorangeProblem(epochs, nav));
 }
 
-std::vector<FGOProcessor::GeometryFreeSlipShadowEpoch>
-FGOProcessor::analyzeGeometryFreeSlipShadow(const FGOProblem& problem,
-                                            double threshold_m,
-                                            double max_gap_s) {
-    struct PhaseSample {
-        double single_difference_phase_m = 0.0;
-        bool has_single_difference_doppler = false;
-        double single_difference_doppler_mps = 0.0;
-        std::set<std::size_t> ambiguity_indices;
-    };
-    struct History {
-        GNSSTime time;
-        double geometry_free_m = 0.0;
-        std::set<std::size_t> first_ambiguity_indices;
-        std::set<std::size_t> second_ambiguity_indices;
-    };
-    struct DopplerHistory {
-        GNSSTime time;
-        double single_difference_phase_m = 0.0;
-        double single_difference_doppler_mps = 0.0;
-        std::set<std::size_t> ambiguity_indices;
-    };
-
-    using CarrierKey = std::pair<SatelliteId, SignalType>;
-    using PairKey = std::tuple<SatelliteId, SignalType, SignalType>;
-    std::vector<std::map<CarrierKey, PhaseSample>> phases_by_epoch(
-        problem.epochs.size());
-    for (const auto& factor : problem.double_difference_carrier_factors) {
-        if (factor.epoch_index >= phases_by_epoch.size()) {
-            continue;
-        }
-        auto& phases = phases_by_epoch[factor.epoch_index];
-        const auto add_phase = [&](const CarrierKey& key, double phase_m,
-                                   const ObservationModelDebug& rover_model,
-                                   const ObservationModelDebug& base_model) {
-            auto& sample = phases[key];
-            sample.single_difference_phase_m = phase_m;
-            if (rover_model.has_doppler_residual &&
-                base_model.has_doppler_residual) {
-                sample.has_single_difference_doppler = true;
-                sample.single_difference_doppler_mps =
-                    rover_model.doppler_residual_mps -
-                    base_model.doppler_residual_mps;
-            }
-            sample.ambiguity_indices.insert(factor.ambiguity_index);
-        };
-        add_phase({factor.satellite, factor.signal},
-                  factor.rover_satellite_model.raw_carrier_m -
-                      factor.base_satellite_model.raw_carrier_m,
-                  factor.rover_satellite_model,
-                  factor.base_satellite_model);
-        // A slip on the DD reference contaminates every target ambiguity in
-        // that group, hence the same factor ambiguity is attached here too.
-        add_phase({factor.reference_satellite, factor.signal},
-                  factor.rover_reference_model.raw_carrier_m -
-                      factor.base_reference_model.raw_carrier_m,
-                  factor.rover_reference_model,
-                  factor.base_reference_model);
-    }
-
-    std::vector<GeometryFreeSlipShadowEpoch> shadow(problem.epochs.size());
-    std::map<PairKey, History> history;
-    std::map<CarrierKey, DopplerHistory> doppler_history;
-    std::set<std::size_t> tainted_ambiguities;
-    const double safe_threshold = std::max(0.0, threshold_m);
-    const double safe_max_gap = std::max(0.0, max_gap_s);
-    for (std::size_t epoch_index = 0; epoch_index < phases_by_epoch.size();
-         ++epoch_index) {
-        std::map<SatelliteId, std::vector<std::pair<SignalType, PhaseSample>>>
-            by_satellite;
-        for (const auto& [key, sample] : phases_by_epoch[epoch_index]) {
-            by_satellite[key.first].push_back({key.second, sample});
-        }
-        const GNSSTime time = problem.epochs[epoch_index].time;
-        std::set<CarrierKey> doppler_events;
-        for (const auto& [key, sample] : phases_by_epoch[epoch_index]) {
-            const auto previous = doppler_history.find(key);
-            if (sample.has_single_difference_doppler &&
-                previous != doppler_history.end()) {
-                const double dt = time - previous->second.time;
-                const bool same_arcs =
-                    previous->second.ambiguity_indices ==
-                    sample.ambiguity_indices;
-                if (same_arcs && dt > 0.0 &&
-                    (safe_max_gap <= 0.0 || dt <= safe_max_gap)) {
-                    const double predicted_delta_m =
-                        0.5 * (previous->second.single_difference_doppler_mps +
-                               sample.single_difference_doppler_mps) * dt;
-                    const double innovation_m = std::abs(
-                        sample.single_difference_phase_m -
-                        previous->second.single_difference_phase_m -
-                        predicted_delta_m);
-                    if (innovation_m > 0.20) {
-                        doppler_events.insert(key);
-                        shadow[epoch_index].doppler_max_innovation_m =
-                            std::max(
-                                shadow[epoch_index].doppler_max_innovation_m,
-                                innovation_m);
-                    }
-                }
-            }
-            if (sample.has_single_difference_doppler) {
-                doppler_history[key] = {
-                    time, sample.single_difference_phase_m,
-                    sample.single_difference_doppler_mps,
-                    sample.ambiguity_indices};
-            } else {
-                doppler_history.erase(key);
-            }
-        }
-        shadow[epoch_index].doppler_event_signals =
-            static_cast<int>(doppler_events.size());
-        for (const auto& [satellite, samples] : by_satellite) {
-            for (std::size_t first = 0; first < samples.size(); ++first) {
-                for (std::size_t second = first + 1; second < samples.size();
-                     ++second) {
-                    const auto& first_sample = samples[first].second;
-                    const auto& second_sample = samples[second].second;
-                    const double geometry_free =
-                        first_sample.single_difference_phase_m -
-                        second_sample.single_difference_phase_m;
-                    const PairKey key{satellite, samples[first].first,
-                                      samples[second].first};
-                    const auto previous = history.find(key);
-                    if (previous != history.end()) {
-                        const double dt = time - previous->second.time;
-                        const bool same_arcs =
-                            previous->second.first_ambiguity_indices ==
-                                first_sample.ambiguity_indices &&
-                            previous->second.second_ambiguity_indices ==
-                                second_sample.ambiguity_indices;
-                        if (same_arcs && dt > 0.0 &&
-                            (safe_max_gap <= 0.0 || dt <= safe_max_gap)) {
-                            const double jump = std::abs(
-                                geometry_free - previous->second.geometry_free_m);
-                            if (jump > safe_threshold) {
-                                ++shadow[epoch_index].event_pairs;
-                                shadow[epoch_index].max_jump_m = std::max(
-                                    shadow[epoch_index].max_jump_m, jump);
-                                shadow[epoch_index].event_satellite_signals.insert(
-                                    {satellite, samples[first].first});
-                                shadow[epoch_index].event_satellite_signals.insert(
-                                    {satellite, samples[second].first});
-                                tainted_ambiguities.insert(
-                                    first_sample.ambiguity_indices.begin(),
-                                    first_sample.ambiguity_indices.end());
-                                tainted_ambiguities.insert(
-                                    second_sample.ambiguity_indices.begin(),
-                                    second_sample.ambiguity_indices.end());
-                                const int doppler_matches =
-                                    static_cast<int>(doppler_events.count(
-                                        {satellite, samples[first].first})) +
-                                    static_cast<int>(doppler_events.count(
-                                        {satellite, samples[second].first}));
-                                if (doppler_matches == 1) {
-                                    ++shadow[epoch_index]
-                                          .doppler_isolated_event_pairs;
-                                }
-                            }
-                        }
-                    }
-                    history[key] = {time, geometry_free,
-                                    first_sample.ambiguity_indices,
-                                    second_sample.ambiguity_indices};
-                }
-            }
-        }
-
-        std::set<std::size_t> tainted_present;
-        for (const auto& [key, sample] : phases_by_epoch[epoch_index]) {
-            (void)key;
-            for (const std::size_t ambiguity_index : sample.ambiguity_indices) {
-                if (tainted_ambiguities.count(ambiguity_index) != 0) {
-                    tainted_present.insert(ambiguity_index);
-                }
-            }
-        }
-        shadow[epoch_index].tainted_ambiguities =
-            static_cast<int>(tainted_present.size());
-    }
-    return shadow;
-}
-
-std::vector<FGOProcessor::SingleDifferenceTdcpFactor>
-FGOProcessor::buildClockResilientTemporalCarrierShadow(
-    const FGOProblem& problem, double sigma_m, double max_gap_s) {
-    using GroupKey = std::pair<GNSSSystem, SignalType>;
-    using ObservationKey = std::pair<SatelliteId, SignalType>;
-    using HistoryKey = std::tuple<SatelliteId, SatelliteId, SignalType>;
-
-    struct HistorySample {
-        std::size_t epoch_index = 0;
-        double residual_m = 0.0;
-        Vector3d los = Vector3d::Zero();
-        int arc_length_epochs = 1;
-        bool has_sd_doppler = false;
-        double sd_doppler_mps = 0.0;
-        double sd_doppler_sigma_mps = 0.0;
-    };
-
-    const auto& observations = problem.carrier_observations.empty()
-        ? problem.double_difference_pseudorange_observations
-        : problem.carrier_observations;
-    std::vector<std::map<ObservationKey, const CarrierPhaseFactor*>> by_epoch(
-        problem.epochs.size());
-    for (const auto& observation : observations) {
-        if (observation.epoch_index >= by_epoch.size() ||
-            !observation.has_carrier_phase || observation.loss_of_lock ||
-            !std::isfinite(observation.corrected_carrier_m) ||
-            !std::isfinite(observation.model_debug.geometric_range_m) ||
-            !observation.los.allFinite()) {
-            continue;
-        }
-        by_epoch[observation.epoch_index]
-                [{observation.satellite, observation.signal}] = &observation;
-    }
-
-    std::vector<SingleDifferenceTdcpFactor> factors;
-    std::map<GroupKey, SatelliteId> causal_references;
-    std::map<HistoryKey, HistorySample> previous;
-    const double safe_sigma = std::max(1e-4, sigma_m);
-    const double safe_max_gap = std::max(0.0, max_gap_s);
-
-    for (std::size_t epoch_index = 0; epoch_index < by_epoch.size();
-         ++epoch_index) {
-        std::map<GroupKey, std::vector<const CarrierPhaseFactor*>> groups;
-        for (const auto& [key, observation] : by_epoch[epoch_index]) {
-            (void)key;
-            groups[{observation->satellite.system, observation->signal}]
-                .push_back(observation);
-        }
-
-        std::map<HistoryKey, HistorySample> current;
-        for (const auto& [group_key, group] : groups) {
-            if (group.size() < 2) {
-                causal_references.erase(group_key);
-                continue;
-            }
-
-            const CarrierPhaseFactor* reference = nullptr;
-            const auto prior_reference = causal_references.find(group_key);
-            if (prior_reference != causal_references.end()) {
-                for (const auto* candidate : group) {
-                    if (candidate->satellite == prior_reference->second) {
-                        reference = candidate;
-                        break;
-                    }
-                }
-            }
-            if (reference == nullptr) {
-                reference = *std::max_element(
-                    group.begin(), group.end(),
-                    [](const auto* lhs, const auto* rhs) {
-                        if (lhs->elevation_rad != rhs->elevation_rad) {
-                            return lhs->elevation_rad < rhs->elevation_rad;
-                        }
-                        return rhs->satellite < lhs->satellite;
-                    });
-            }
-            causal_references[group_key] = reference->satellite;
-
-            const double reference_residual =
-                reference->corrected_carrier_m -
-                reference->model_debug.geometric_range_m;
-            for (const auto* observation : group) {
-                if (observation->satellite == reference->satellite) {
-                    continue;
-                }
-                const double residual =
-                    observation->corrected_carrier_m -
-                    observation->model_debug.geometric_range_m -
-                    reference_residual;
-                const Vector3d los = observation->los - reference->los;
-                const HistoryKey history_key{observation->satellite,
-                                             reference->satellite,
-                                             observation->signal};
-                const bool has_sd_doppler =
-                    observation->has_doppler_residual &&
-                    reference->has_doppler_residual;
-                const double sd_doppler_mps = has_sd_doppler
-                    ? observation->doppler_residual_mps -
-                          reference->doppler_residual_mps
-                    : 0.0;
-                const double sd_doppler_sigma_mps = has_sd_doppler
-                    ? std::hypot(observation->doppler_sigma_mps,
-                                 reference->doppler_sigma_mps)
-                    : 0.0;
-                current[history_key] = {epoch_index, residual, los, 1,
-                                        has_sd_doppler, sd_doppler_mps,
-                                        sd_doppler_sigma_mps};
-
-                const auto prior = previous.find(history_key);
-                if (prior == previous.end()) {
-                    continue;
-                }
-                const double dt = problem.epochs[epoch_index].time -
-                                  problem.epochs[prior->second.epoch_index].time;
-                if (prior->second.epoch_index + 1 != epoch_index || dt <= 0.0 ||
-                    (safe_max_gap > 0.0 && dt > safe_max_gap)) {
-                    continue;
-                }
-
-                const double delta = residual - prior->second.residual_m;
-                if (!std::isfinite(delta)) {
-                    continue;
-                }
-                SingleDifferenceTdcpFactor factor;
-                factor.previous_epoch_index = prior->second.epoch_index;
-                factor.current_epoch_index = epoch_index;
-                factor.satellite = observation->satellite;
-                factor.reference_satellite = reference->satellite;
-                factor.signal = observation->signal;
-                factor.previous_los = prior->second.los;
-                factor.los = los;
-                factor.delta_carrier_m = delta;
-                factor.target_ambiguity_index = observation->ambiguity_index;
-                factor.reference_ambiguity_index = reference->ambiguity_index;
-                factor.arc_length_epochs = prior->second.arc_length_epochs + 1;
-                factor.dt_s = dt;
-                factor.has_doppler_witness =
-                    prior->second.has_sd_doppler && has_sd_doppler;
-                factor.previous_sd_doppler_mps =
-                    prior->second.sd_doppler_mps;
-                factor.current_sd_doppler_mps = sd_doppler_mps;
-                factor.previous_sd_doppler_sigma_mps =
-                    prior->second.sd_doppler_sigma_mps;
-                factor.current_sd_doppler_sigma_mps =
-                    sd_doppler_sigma_mps;
-                current[history_key].arc_length_epochs = factor.arc_length_epochs;
-                const double sin_el = std::max(
-                    0.1, std::sin(std::min(observation->elevation_rad,
-                                           reference->elevation_rad)));
-                factor.sigma_m = safe_sigma / std::sqrt(sin_el);
-                factor.elevation_rad =
-                    std::min(observation->elevation_rad,
-                             reference->elevation_rad);
-                factors.push_back(factor);
-            }
-        }
-        previous = std::move(current);
-    }
-    return factors;
-}
-
-std::vector<FGOProcessor::TemporalCarrierShadowFactorDiagnostics>
-FGOProcessor::classifyClockResilientTemporalCarrierShadow(
-    const FGOProblem& problem,
-    const std::vector<SingleDifferenceTdcpFactor>& factors,
-    const std::vector<Vector3d>& epoch_positions_ecef,
-    std::vector<FGOEpochDiagnostics>& epoch_diagnostics,
-    const std::vector<GeometryFreeSlipShadowEpoch>* geometry_free_shadow,
-    const std::vector<std::set<std::size_t>>*
-        fde_rejected_ambiguities_by_epoch) {
-    const std::size_t num_epochs = problem.epochs.size();
-    if (epoch_diagnostics.size() < num_epochs) {
-        epoch_diagnostics.resize(num_epochs);
-    }
-    if (epoch_positions_ecef.size() < num_epochs) {
-        return {};
-    }
-
-    using DopplerKey =
-        std::tuple<std::size_t, SatelliteId, SatelliteId, SignalType>;
-    std::map<DopplerKey, SingleDifferenceDopplerFactor> sd_doppler_by_factor;
-    for (const auto& doppler : problem.single_difference_doppler_factors) {
-        sd_doppler_by_factor[{doppler.epoch_index, doppler.satellite,
-                              doppler.reference_satellite, doppler.signal}] =
-            doppler;
-    }
-
-    constexpr double kResidualOutlierThresholdM = 0.10;
-    constexpr double kDopplerNormalizedInnovationThreshold = 5.0;
-    constexpr std::size_t kDopplerCalibrationWarmup = 10;
-    constexpr std::size_t kDopplerCalibrationWindow = 30;
-    constexpr double kDopplerCalibratedScoreThreshold = 5.0;
-    using DopplerCalibrationKey =
-        std::tuple<std::size_t, SatelliteId, SatelliteId, SignalType>;
-    std::map<DopplerCalibrationKey, std::vector<double>>
-        doppler_innovation_history;
-    const auto median = [](std::vector<double> values) {
-        const std::size_t middle = values.size() / 2;
-        std::nth_element(values.begin(), values.begin() + middle, values.end());
-        double value = values[middle];
-        if (values.size() % 2 == 0) {
-            const auto lower =
-                std::max_element(values.begin(), values.begin() + middle);
-            value = 0.5 * (value + *lower);
-        }
-        return value;
-    };
-
-    std::vector<double> sum_sq(num_epochs, 0.0);
-    std::vector<TemporalCarrierShadowFactorDiagnostics> classified_factors;
-    classified_factors.reserve(factors.size());
-    for (const auto& factor : factors) {
-        if (factor.previous_epoch_index >= num_epochs ||
-            factor.current_epoch_index >= num_epochs) {
-            continue;
-        }
-        const Vector3d previous_delta =
-            epoch_positions_ecef[factor.previous_epoch_index] -
-            problem.epochs[factor.previous_epoch_index].position_ecef;
-        const Vector3d current_delta =
-            epoch_positions_ecef[factor.current_epoch_index] -
-            problem.epochs[factor.current_epoch_index].position_ecef;
-        const double predicted = factor.los.dot(current_delta) -
-                                 factor.previous_los.dot(previous_delta);
-        const double residual = factor.delta_carrier_m - predicted;
-        if (!std::isfinite(residual)) {
-            continue;
-        }
-        auto& diagnostics = epoch_diagnostics[factor.current_epoch_index];
-        ++diagnostics.clock_resilient_tdcp_factors;
-        sum_sq[factor.current_epoch_index] += residual * residual;
-        diagnostics.clock_resilient_tdcp_max_abs_m =
-            std::max(diagnostics.clock_resilient_tdcp_max_abs_m,
-                     std::abs(residual));
-
-        TemporalCarrierShadowFactorDiagnostics classified;
-        classified.factor = factor;
-        classified.residual_m = residual;
-        classified.normalized_residual =
-            residual / std::max(1e-4, factor.sigma_m);
-        classified.residual_outlier =
-            std::abs(residual) > kResidualOutlierThresholdM;
-
-        const auto evaluate_doppler = [&](double previous_mps,
-                                          double current_mps,
-                                          double previous_sigma_mps,
-                                          double current_sigma_mps,
-                                          double dt) {
-            if (dt <= 0.0 || !std::isfinite(dt) ||
-                previous_sigma_mps <= 0.0 || current_sigma_mps <= 0.0) {
-                return;
-            }
-            const double predicted_delta =
-                0.5 * (previous_mps + current_mps) * dt;
-            classified.doppler_innovation_signed_m =
-                factor.delta_carrier_m - predicted_delta;
-            classified.doppler_innovation_m =
-                std::abs(classified.doppler_innovation_signed_m);
-            const double integrated_doppler_sigma_m =
-                0.5 * dt *
-                std::hypot(previous_sigma_mps, current_sigma_mps);
-            classified.doppler_innovation_sigma_m =
-                std::hypot(factor.sigma_m, integrated_doppler_sigma_m);
-            classified.normalized_doppler_innovation =
-                classified.doppler_innovation_m /
-                std::max(1e-4, classified.doppler_innovation_sigma_m);
-            classified.doppler_evaluated =
-                std::isfinite(classified.doppler_innovation_m) &&
-                std::isfinite(classified.normalized_doppler_innovation);
-            classified.doppler_outlier =
-                classified.doppler_evaluated &&
-                classified.normalized_doppler_innovation >
-                    kDopplerNormalizedInnovationThreshold;
-        };
-        if (factor.has_doppler_witness) {
-            evaluate_doppler(factor.previous_sd_doppler_mps,
-                             factor.current_sd_doppler_mps,
-                             factor.previous_sd_doppler_sigma_mps,
-                             factor.current_sd_doppler_sigma_mps, factor.dt_s);
-        } else {
-            const DopplerKey previous_key{
-                factor.previous_epoch_index, factor.satellite,
-                factor.reference_satellite, factor.signal};
-            const DopplerKey current_key{
-                factor.current_epoch_index, factor.satellite,
-                factor.reference_satellite, factor.signal};
-            const auto previous_doppler = sd_doppler_by_factor.find(previous_key);
-            const auto current_doppler = sd_doppler_by_factor.find(current_key);
-            if (previous_doppler != sd_doppler_by_factor.end() &&
-                current_doppler != sd_doppler_by_factor.end()) {
-                const double dt =
-                    problem.epochs[factor.current_epoch_index].time -
-                    problem.epochs[factor.previous_epoch_index].time;
-                evaluate_doppler(previous_doppler->second.residual_mps,
-                                 current_doppler->second.residual_mps,
-                                 previous_doppler->second.sigma_mps,
-                                 current_doppler->second.sigma_mps, dt);
-            }
-        }
-
-        if (classified.doppler_evaluated && factor.arc_length_epochs >= 2) {
-            const std::size_t arc_start_epoch =
-                factor.current_epoch_index + 1 -
-                static_cast<std::size_t>(factor.arc_length_epochs);
-            const DopplerCalibrationKey key{
-                arc_start_epoch, factor.satellite,
-                factor.reference_satellite, factor.signal};
-            auto& history = doppler_innovation_history[key];
-            if (history.size() >= kDopplerCalibrationWarmup) {
-                classified.doppler_bias_m = median(history);
-                std::vector<double> deviations;
-                deviations.reserve(history.size());
-                for (const double sample : history) {
-                    deviations.push_back(
-                        std::abs(sample - classified.doppler_bias_m));
-                }
-                const double robust_sigma_m = 1.4826 * median(deviations);
-                classified.doppler_calibrated_scale_m = std::max(
-                    {1e-4, classified.doppler_innovation_sigma_m,
-                     robust_sigma_m});
-                classified.doppler_centered_innovation_m =
-                    classified.doppler_innovation_signed_m -
-                    classified.doppler_bias_m;
-                classified.doppler_calibrated_score =
-                    std::abs(classified.doppler_centered_innovation_m) /
-                    classified.doppler_calibrated_scale_m;
-                classified.doppler_calibration_evaluated =
-                    std::isfinite(classified.doppler_calibrated_score);
-                classified.doppler_calibrated_outlier =
-                    classified.doppler_calibration_evaluated &&
-                    classified.doppler_calibrated_score >
-                        kDopplerCalibratedScoreThreshold;
-            }
-            history.push_back(classified.doppler_innovation_signed_m);
-            if (history.size() > kDopplerCalibrationWindow) {
-                history.erase(history.begin());
-            }
-        }
-
-        if (geometry_free_shadow != nullptr &&
-            factor.current_epoch_index < geometry_free_shadow->size()) {
-            const auto& events =
-                (*geometry_free_shadow)[factor.current_epoch_index]
-                    .event_satellite_signals;
-            classified.geometry_free_witness =
-                events.count({factor.satellite, factor.signal}) != 0 ||
-                events.count({factor.reference_satellite, factor.signal}) != 0;
-        }
-        classified.carrier_hold_witness = diagnostics.carrier_hold_active;
-        if (fde_rejected_ambiguities_by_epoch != nullptr &&
-            factor.current_epoch_index <
-                fde_rejected_ambiguities_by_epoch->size()) {
-            const auto& rejected = (*fde_rejected_ambiguities_by_epoch)
-                [factor.current_epoch_index];
-            classified.carrier_fde_witness =
-                rejected.count(factor.target_ambiguity_index) != 0 ||
-                rejected.count(factor.reference_ambiguity_index) != 0;
-        }
-
-        const bool witnessed = classified.doppler_calibrated_outlier ||
-            classified.geometry_free_witness ||
-            classified.carrier_hold_witness ||
-            classified.carrier_fde_witness;
-        if (!classified.residual_outlier) {
-            classified.classification =
-                TemporalCarrierShadowClassification::Clean;
-            ++diagnostics.clock_resilient_tdcp_clean;
-        } else if (witnessed) {
-            classified.classification =
-                TemporalCarrierShadowClassification::WitnessedOutlier;
-            ++diagnostics.clock_resilient_tdcp_witnessed_outliers;
-        } else {
-            classified.classification =
-                TemporalCarrierShadowClassification::UnexplainedOutlier;
-            ++diagnostics.clock_resilient_tdcp_unexplained_outliers;
-        }
-        classified_factors.push_back(std::move(classified));
-    }
-    for (std::size_t i = 0; i < num_epochs; ++i) {
-        const int count = epoch_diagnostics[i].clock_resilient_tdcp_factors;
-        if (count > 0) {
-            epoch_diagnostics[i].clock_resilient_tdcp_rms_m =
-                std::sqrt(sum_sq[i] / static_cast<double>(count));
-        }
-    }
-    return classified_factors;
-}
-
-std::vector<FGOProcessor::PredictedDdprQualityFactorDiagnostics>
-FGOProcessor::analyzePredictedDdprQualityShadow(
-    const FGOProblem& problem,
-    const std::vector<Vector3d>& previous_solution_positions_ecef,
-    const std::vector<Vector3d>& predicted_positions_ecef,
-    double doppler_sigma_mps,
-    double normalized_outlier_threshold,
-    double max_gap_s) {
-    using Key = std::tuple<SatelliteId, SatelliteId, SignalType>;
-    struct History {
-        const DoubleDifferencePseudorangeFactor* factor = nullptr;
-        std::size_t epoch_index = 0;
-        double dd_doppler_mps = 0.0;
-        bool has_dd_doppler = false;
-        int pair_age_epochs = 0;
-    };
-
-    const auto finite_position = [](const Vector3d& position) {
-        return position.allFinite() && position.norm() > 0.0;
-    };
-    const auto dd_range = [](const DoubleDifferencePseudorangeFactor& factor,
-                             const Vector3d& rover_position_ecef) {
-        const double rover_target =
-            (factor.rover_satellite_position_ecef - rover_position_ecef).norm();
-        const double rover_reference =
-            (factor.rover_reference_position_ecef - rover_position_ecef).norm();
-        const double base_target =
-            (factor.base_satellite_position_ecef - factor.base_position_ecef)
-                .norm();
-        const double base_reference =
-            (factor.base_reference_position_ecef - factor.base_position_ecef)
-                .norm();
-        return (rover_target - base_target) -
-               (rover_reference - base_reference);
-    };
-    const auto dd_doppler = [](const DoubleDifferencePseudorangeFactor& factor,
-                               double& value) {
-        const auto& rt = factor.rover_satellite_model;
-        const auto& rr = factor.rover_reference_model;
-        const auto& bt = factor.base_satellite_model;
-        const auto& br = factor.base_reference_model;
-        if (!rt.has_doppler_residual || !rr.has_doppler_residual ||
-            !bt.has_doppler_residual || !br.has_doppler_residual) {
-            return false;
-        }
-        value = (rt.doppler_residual_mps - bt.doppler_residual_mps) -
-                (rr.doppler_residual_mps - br.doppler_residual_mps);
-        return std::isfinite(value);
-    };
-
-    const std::size_t num_epochs = problem.epochs.size();
-    std::vector<std::vector<const DoubleDifferencePseudorangeFactor*>>
-        factors_by_epoch(num_epochs);
-    for (const auto& factor : problem.double_difference_pseudorange_factors) {
-        if (factor.epoch_index < num_epochs) {
-            factors_by_epoch[factor.epoch_index].push_back(&factor);
-        }
-    }
-
-    const double safe_doppler_sigma = std::max(1e-6, doppler_sigma_mps);
-    const double safe_threshold = std::max(0.0, normalized_outlier_threshold);
-    const double safe_max_gap = std::max(0.0, max_gap_s);
-    std::map<Key, History> history;
-    std::vector<PredictedDdprQualityFactorDiagnostics> result;
-    result.reserve(problem.double_difference_pseudorange_factors.size());
-    for (std::size_t epoch_index = 0; epoch_index < num_epochs; ++epoch_index) {
-        std::map<Key, History> current;
-        for (const auto* factor : factors_by_epoch[epoch_index]) {
-            const Key key{factor->satellite, factor->reference_satellite,
-                          factor->signal};
-            double current_dd_doppler = 0.0;
-            const bool current_has_doppler =
-                dd_doppler(*factor, current_dd_doppler);
-            History current_history{factor, epoch_index, current_dd_doppler,
-                                    current_has_doppler, 1};
-            const auto prior = history.find(key);
-            if (prior == history.end()) {
-                current[key] = current_history;
-                continue;
-            }
-
-            const std::size_t previous_epoch_index = prior->second.epoch_index;
-            const double dt = problem.epochs[epoch_index].time -
-                              problem.epochs[previous_epoch_index].time;
-            if (previous_epoch_index + 1 != epoch_index || dt <= 0.0 ||
-                (safe_max_gap > 0.0 && dt > safe_max_gap)) {
-                current[key] = current_history;
-                continue;
-            }
-
-            current_history.pair_age_epochs = prior->second.pair_age_epochs + 1;
-            current[key] = current_history;
-            PredictedDdprQualityFactorDiagnostics diagnostic;
-            diagnostic.previous_epoch_index = previous_epoch_index;
-            diagnostic.current_epoch_index = epoch_index;
-            diagnostic.satellite = factor->satellite;
-            diagnostic.reference_satellite = factor->reference_satellite;
-            diagnostic.signal = factor->signal;
-            diagnostic.dt_s = dt;
-            diagnostic.pair_age_epochs = current_history.pair_age_epochs;
-            diagnostic.measured_ddpr_change_m =
-                factor->observed_dd_pseudorange_m -
-                prior->second.factor->observed_dd_pseudorange_m;
-            diagnostic.elevation_rad = std::min(
-                factor->rover_satellite_model.elevation_rad,
-                factor->rover_reference_model.elevation_rad);
-            diagnostic.target_snr_dbhz =
-                factor->rover_satellite_model.snr_dbhz;
-            diagnostic.reference_snr_dbhz =
-                factor->rover_reference_model.snr_dbhz;
-
-            if (current_has_doppler && prior->second.has_dd_doppler) {
-                diagnostic.doppler_evaluated = true;
-                diagnostic.doppler_predicted_change_m =
-                    0.5 * (prior->second.dd_doppler_mps +
-                           current_dd_doppler) *
-                    dt;
-                diagnostic.doppler_innovation_m =
-                    diagnostic.measured_ddpr_change_m -
-                    diagnostic.doppler_predicted_change_m;
-                const double measured_sigma = std::hypot(
-                    prior->second.factor->sigma_m, factor->sigma_m);
-                // Four receiver/satellite Doppler terms form each DD. The
-                // trapezoidal difference uses two epochs, yielding
-                // sqrt(2)*dt*sigma for equal independent link sigmas.
-                const double integrated_doppler_sigma =
-                    std::sqrt(2.0) * dt * safe_doppler_sigma;
-                diagnostic.doppler_innovation_sigma_m =
-                    std::hypot(measured_sigma, integrated_doppler_sigma);
-                diagnostic.normalized_doppler_innovation =
-                    std::abs(diagnostic.doppler_innovation_m) /
-                    diagnostic.doppler_innovation_sigma_m;
-            }
-
-            if (previous_epoch_index <
-                    previous_solution_positions_ecef.size() &&
-                epoch_index < predicted_positions_ecef.size() &&
-                finite_position(previous_solution_positions_ecef[
-                    previous_epoch_index]) &&
-                finite_position(predicted_positions_ecef[epoch_index])) {
-                diagnostic.imu_geometry_evaluated = true;
-                diagnostic.imu_predicted_change_m =
-                    dd_range(*factor, predicted_positions_ecef[epoch_index]) -
-                    dd_range(*prior->second.factor,
-                             previous_solution_positions_ecef[
-                                 previous_epoch_index]);
-                diagnostic.previous_predicted_ddpr_residual_m =
-                    prior->second.factor->observed_dd_pseudorange_m -
-                    dd_range(*prior->second.factor,
-                             previous_solution_positions_ecef[
-                                 previous_epoch_index]);
-                diagnostic.current_predicted_ddpr_residual_m =
-                    factor->observed_dd_pseudorange_m -
-                    dd_range(*factor,
-                             predicted_positions_ecef[epoch_index]);
-                diagnostic.imu_innovation_m =
-                    diagnostic.measured_ddpr_change_m -
-                    diagnostic.imu_predicted_change_m;
-                diagnostic.imu_innovation_sigma_m = std::hypot(
-                    prior->second.factor->sigma_m, factor->sigma_m);
-                diagnostic.normalized_imu_innovation =
-                    std::abs(diagnostic.imu_innovation_m) /
-                    std::max(1e-6, diagnostic.imu_innovation_sigma_m);
-            }
-
-            if (!diagnostic.doppler_evaluated &&
-                !diagnostic.imu_geometry_evaluated) {
-                diagnostic.proposed_action =
-                    PredictedDdprQualityAction::Unavailable;
-            } else {
-                diagnostic.proposed_action = PredictedDdprQualityAction::Keep;
-                if (diagnostic.doppler_evaluated &&
-                    diagnostic.imu_geometry_evaluated &&
-                    diagnostic.normalized_doppler_innovation > safe_threshold &&
-                    diagnostic.normalized_imu_innovation > safe_threshold) {
-                    diagnostic.proposed_action =
-                        PredictedDdprQualityAction::Downweight;
-                }
-            }
-            result.push_back(diagnostic);
-        }
-        history = std::move(current);
-    }
-    return result;
-}
-
-std::vector<FGOProcessor::PredictedDdprBiasStateDiagnostics>
-FGOProcessor::analyzePredictedDdprBiasStateShadow(
-    const std::vector<PredictedDdprQualityFactorDiagnostics>& quality_rows,
-    double process_noise_m_sqrt_s,
-    double initial_sigma_m,
-    double min_measurement_sigma_m,
-    double robust_update_sigma,
-    int min_prior_updates) {
-    using Key = std::tuple<SatelliteId, SatelliteId, SignalType>;
-    struct State {
-        double mean_m = 0.0;
-        double variance_m2 = 0.0;
-        int updates = 0;
-        std::size_t last_epoch_index = 0;
-        bool initialized = false;
-    };
-
-    const double q = std::max(0.0, process_noise_m_sqrt_s);
-    const double initial_sigma = std::max(1e-6, initial_sigma_m);
-    const double min_measurement_sigma =
-        std::max(1e-6, min_measurement_sigma_m);
-    const double clip_sigma = std::max(0.0, robust_update_sigma);
-    const int burn_in = std::max(0, min_prior_updates);
-    std::map<Key, State> states;
-    std::vector<PredictedDdprBiasStateDiagnostics> result;
-    result.reserve(quality_rows.size());
-
-    for (const auto& row : quality_rows) {
-        PredictedDdprBiasStateDiagnostics diagnostic;
-        diagnostic.previous_epoch_index = row.previous_epoch_index;
-        diagnostic.current_epoch_index = row.current_epoch_index;
-        diagnostic.satellite = row.satellite;
-        diagnostic.reference_satellite = row.reference_satellite;
-        diagnostic.signal = row.signal;
-        diagnostic.dt_s = row.dt_s;
-        diagnostic.pair_age_epochs = row.pair_age_epochs;
-        diagnostic.raw_residual_m =
-            row.current_predicted_ddpr_residual_m;
-
-        const Key key{row.satellite, row.reference_satellite, row.signal};
-        auto& state = states[key];
-        const bool continuous = state.initialized &&
-                                state.last_epoch_index ==
-                                    row.previous_epoch_index &&
-                                row.current_epoch_index ==
-                                    row.previous_epoch_index + 1 &&
-                                std::isfinite(row.dt_s) && row.dt_s > 0.0;
-        if (!continuous) {
-            state = State{};
-            state.variance_m2 = initial_sigma * initial_sigma;
-            diagnostic.continuity_reset = true;
-        } else {
-            state.variance_m2 += q * q * row.dt_s;
-        }
-
-        diagnostic.prior_updates = state.updates;
-        diagnostic.prior_bias_m = state.mean_m;
-        diagnostic.prior_sigma_m =
-            std::sqrt(std::max(0.0, state.variance_m2));
-        diagnostic.corrected_residual_m =
-            diagnostic.raw_residual_m - diagnostic.prior_bias_m;
-        diagnostic.prediction_usable =
-            continuous && state.updates >= burn_in &&
-            row.imu_geometry_evaluated &&
-            std::isfinite(diagnostic.raw_residual_m);
-
-        // The quality monitor propagates the two adjacent DDPR sigmas with
-        // hypot(). Divide by sqrt(2) to obtain an equal-row approximation;
-        // the floor deliberately prevents an overconfident shadow state.
-        const double approximated_row_sigma =
-            row.imu_innovation_sigma_m / std::sqrt(2.0);
-        diagnostic.measurement_sigma_m = std::max(
-            min_measurement_sigma,
-            std::isfinite(approximated_row_sigma)
-                ? approximated_row_sigma
-                : min_measurement_sigma);
-
-        const bool update_valid =
-            row.imu_geometry_evaluated &&
-            std::isfinite(diagnostic.raw_residual_m) &&
-            std::isfinite(state.mean_m) &&
-            std::isfinite(state.variance_m2);
-        if (update_valid) {
-            const double measurement_variance =
-                diagnostic.measurement_sigma_m *
-                diagnostic.measurement_sigma_m;
-            const double innovation_m =
-                diagnostic.raw_residual_m - state.mean_m;
-            diagnostic.innovation_sigma_m = std::sqrt(std::max(
-                1e-12, state.variance_m2 + measurement_variance));
-            diagnostic.normalized_innovation =
-                std::abs(innovation_m) / diagnostic.innovation_sigma_m;
-            diagnostic.applied_innovation_m = innovation_m;
-            if (clip_sigma > 0.0) {
-                const double limit =
-                    clip_sigma * diagnostic.innovation_sigma_m;
-                if (std::abs(diagnostic.applied_innovation_m) > limit) {
-                    diagnostic.applied_innovation_m =
-                        std::copysign(limit,
-                                      diagnostic.applied_innovation_m);
-                    diagnostic.update_clipped = true;
-                }
-            }
-            const double kalman_gain =
-                state.variance_m2 /
-                (state.variance_m2 + measurement_variance);
-            state.mean_m +=
-                kalman_gain * diagnostic.applied_innovation_m;
-            state.variance_m2 =
-                std::max(0.0, (1.0 - kalman_gain) * state.variance_m2);
-            ++state.updates;
-            state.last_epoch_index = row.current_epoch_index;
-            state.initialized = true;
-            diagnostic.update_applied = true;
-        } else {
-            // Invalid geometry breaks the causal chain. A later valid row
-            // starts from the configured prior instead of stale state.
-            states.erase(key);
-        }
-
-        if (diagnostic.update_applied) {
-            diagnostic.posterior_bias_m = state.mean_m;
-            diagnostic.posterior_sigma_m =
-                std::sqrt(std::max(0.0, state.variance_m2));
-        }
-        result.push_back(diagnostic);
-    }
-    return result;
-}
 
 FGOProcessor::FGOResult FGOProcessor::optimize(
     const std::vector<ObservationData>& rover_epochs,
